+++ static/js/webflow.js
@@ -5,2563 +5,2065 @@
  *   var Webflow = Webflow || [];
  *   Webflow.push(readyFunction);
  */
-(function () {
-    'use strict';
+! function() {
+    "use strict";
 
-    function __commonjs(fn, module) {
-        return module = {exports: {}}, fn(module, module.exports), module.exports;
+    function t(t, e) {
+        return e = {
+            exports: {}
+        }, t(e, e.exports), e.exports
     }
-
-    /*!
-     * tram.js v0.8.2-global
-     * Cross-browser CSS3 transitions in JavaScript
-     * https://github.com/bkwld/tram
-     * MIT License
-     */
-    window.tram = function (a) {
-        function b(a, b) {
-            var c = new M.Bare;
-            return c.init(a, b)
+    window.tram = function(t) {
+        function e(t, e) {
+            return (new S.Bare).init(t, e)
         }
 
-        function c(a) {
-            return a.replace(/[A-Z]/g, function (a) {
-                return "-" + a.toLowerCase()
+        function n(t) {
+            return t.replace(/[A-Z]/g, function(t) {
+                return "-" + t.toLowerCase()
             })
         }
 
-        function d(a) {
-            var b = parseInt(a.slice(1), 16), c = b >> 16 & 255, d = b >> 8 & 255, e = 255 & b;
-            return [c, d, e]
+        function i(t) {
+            var e = parseInt(t.slice(1), 16);
+            return [e >> 16 & 255, e >> 8 & 255, 255 & e]
         }
 
-        function e(a, b, c) {
-            return "#" + (1 << 24 | a << 16 | b << 8 | c).toString(16).slice(1)
+        function r(t, e, n) {
+            return "#" + (1 << 24 | t << 16 | e << 8 | n).toString(16).slice(1)
         }
 
-        function f() {
+        function o() {}
+
+        function a(t, e) {
+            c("Type warning: Expected: [" + t + "] Got: [" + typeof e + "] " + e)
         }
 
-        function g(a, b) {
-            j("Type warning: Expected: [" + a + "] Got: [" + typeof b + "] " + b)
+        function s(t, e, n) {
+            c("Units do not match [" + t + "]: " + e + ", " + n)
         }
 
-        function h(a, b, c) {
-            j("Units do not match [" + a + "]: " + b + ", " + c)
+        function u(t, e, n) {
+            if (void 0 !== e && (n = e), void 0 === t) return n;
+            var i = n;
+            return V.test(t) || !K.test(t) ? i = parseInt(t, 10) : K.test(t) && (i = 1e3 * parseFloat(t)), 0 > i &&
+                (i = 0), i === i ? i : n
         }
 
-        function i(a, b, c) {
-            if (void 0 !== b && (c = b), void 0 === a)return c;
-            var d = c;
-            return $.test(a) || !_.test(a) ? d = parseInt(a, 10) : _.test(a) && (d = 1e3 * parseFloat(a)), 0 > d && (d = 0), d === d ? d : c
+        function c(t) {
+            Z.debug && window && window.console.warn(t)
         }
 
-        function j(a) {
-            U.debug && window && window.console.warn(a)
-        }
-
-        function k(a) {
-            for (var b = -1, c = a ? a.length : 0, d = []; ++b < c;) {
-                var e = a[b];
-                e && d.push(e)
+        function l(t) {
+            for (var e = -1, n = t ? t.length : 0, i = []; ++e < n;) {
+                var r = t[e];
+                r && i.push(r)
             }
-            return d
+            return i
         }
-
-        var l = function (a, b, c) {
-                function d(a) {
-                    return "object" == typeof a
+        var f = function(t, e, n) {
+                function i(t) {
+                    return "object" == typeof t
                 }
 
-                function e(a) {
-                    return "function" == typeof a
+                function r(t) {
+                    return "function" == typeof t
                 }
 
-                function f() {
-                }
+                function o() {}
 
-                function g(h, i) {
-                    function j() {
-                        var a = new k;
-                        return e(a.init) && a.init.apply(a, arguments), a
+                function a(s, u) {
+                    function c() {
+                        var t = new l;
+                        return r(t.init) && t.init.apply(t, arguments), t
                     }
 
-                    function k() {
-                    }
-
-                    i === c && (i = h, h = Object), j.Bare = k;
-                    var l, m = f[a] = h[a], n = k[a] = j[a] = new f;
-                    return n.constructor = j, j.mixin = function (b) {
-                        return k[a] = j[a] = g(j, b)[a], j
-                    }, j.open = function (a) {
-                        if (l = {}, e(a) ? l = a.call(j, n, m, j, h) : d(a) && (l = a), d(l))for (var c in l)b.call(l, c) && (n[c] = l[c]);
-                        return e(n.init) || (n.init = h), j
-                    }, j.open(i)
+                    function l() {}
+                    u === n && (u = s, s = Object), c.Bare = l;
+                    var f, h = o[t] = s[t],
+                        d = l[t] = c[t] = new o;
+                    return d.constructor = c, c.mixin = function(e) {
+                        return l[t] = c[t] = a(c, e)[t], c
+                    }, c.open = function(t) {
+                        if (f = {}, r(t) ? f = t.call(c, d, h, c, s) : i(t) && (f = t), i(f))
+                            for (var n in f) e.call(f, n) && (d[n] = f[n]);
+                        return r(d.init) || (d.init = s), c
+                    }, c.open(u)
                 }
-
-                return g
-            }("prototype", {}.hasOwnProperty), m = {
-                ease: ["ease", function (a, b, c, d) {
-                    var e = (a /= d) * a, f = e * a;
-                    return b + c * (-2.75 * f * e + 11 * e * e + -15.5 * f + 8 * e + .25 * a)
-                }], "ease-in": ["ease-in", function (a, b, c, d) {
-                    var e = (a /= d) * a, f = e * a;
-                    return b + c * (-1 * f * e + 3 * e * e + -3 * f + 2 * e)
-                }], "ease-out": ["ease-out", function (a, b, c, d) {
-                    var e = (a /= d) * a, f = e * a;
-                    return b + c * (.3 * f * e + -1.6 * e * e + 2.2 * f + -1.8 * e + 1.9 * a)
-                }], "ease-in-out": ["ease-in-out", function (a, b, c, d) {
-                    var e = (a /= d) * a, f = e * a;
-                    return b + c * (2 * f * e + -5 * e * e + 2 * f + 2 * e)
-                }], linear: ["linear", function (a, b, c, d) {
-                    return c * a / d + b
-                }], "ease-in-quad": ["cubic-bezier(0.550, 0.085, 0.680, 0.530)", function (a, b, c, d) {
-                    return c * (a /= d) * a + b
-                }], "ease-out-quad": ["cubic-bezier(0.250, 0.460, 0.450, 0.940)", function (a, b, c, d) {
-                    return -c * (a /= d) * (a - 2) + b
-                }], "ease-in-out-quad": ["cubic-bezier(0.455, 0.030, 0.515, 0.955)", function (a, b, c, d) {
-                    return (a /= d / 2) < 1 ? c / 2 * a * a + b : -c / 2 * (--a * (a - 2) - 1) + b
-                }], "ease-in-cubic": ["cubic-bezier(0.550, 0.055, 0.675, 0.190)", function (a, b, c, d) {
-                    return c * (a /= d) * a * a + b
-                }], "ease-out-cubic": ["cubic-bezier(0.215, 0.610, 0.355, 1)", function (a, b, c, d) {
-                    return c * ((a = a / d - 1) * a * a + 1) + b
-                }], "ease-in-out-cubic": ["cubic-bezier(0.645, 0.045, 0.355, 1)", function (a, b, c, d) {
-                    return (a /= d / 2) < 1 ? c / 2 * a * a * a + b : c / 2 * ((a -= 2) * a * a + 2) + b
-                }], "ease-in-quart": ["cubic-bezier(0.895, 0.030, 0.685, 0.220)", function (a, b, c, d) {
-                    return c * (a /= d) * a * a * a + b
-                }], "ease-out-quart": ["cubic-bezier(0.165, 0.840, 0.440, 1)", function (a, b, c, d) {
-                    return -c * ((a = a / d - 1) * a * a * a - 1) + b
-                }], "ease-in-out-quart": ["cubic-bezier(0.770, 0, 0.175, 1)", function (a, b, c, d) {
-                    return (a /= d / 2) < 1 ? c / 2 * a * a * a * a + b : -c / 2 * ((a -= 2) * a * a * a - 2) + b
-                }], "ease-in-quint": ["cubic-bezier(0.755, 0.050, 0.855, 0.060)", function (a, b, c, d) {
-                    return c * (a /= d) * a * a * a * a + b
-                }], "ease-out-quint": ["cubic-bezier(0.230, 1, 0.320, 1)", function (a, b, c, d) {
-                    return c * ((a = a / d - 1) * a * a * a * a + 1) + b
-                }], "ease-in-out-quint": ["cubic-bezier(0.860, 0, 0.070, 1)", function (a, b, c, d) {
-                    return (a /= d / 2) < 1 ? c / 2 * a * a * a * a * a + b : c / 2 * ((a -= 2) * a * a * a * a + 2) + b
-                }], "ease-in-sine": ["cubic-bezier(0.470, 0, 0.745, 0.715)", function (a, b, c, d) {
-                    return -c * Math.cos(a / d * (Math.PI / 2)) + c + b
-                }], "ease-out-sine": ["cubic-bezier(0.390, 0.575, 0.565, 1)", function (a, b, c, d) {
-                    return c * Math.sin(a / d * (Math.PI / 2)) + b
-                }], "ease-in-out-sine": ["cubic-bezier(0.445, 0.050, 0.550, 0.950)", function (a, b, c, d) {
-                    return -c / 2 * (Math.cos(Math.PI * a / d) - 1) + b
-                }], "ease-in-expo": ["cubic-bezier(0.950, 0.050, 0.795, 0.035)", function (a, b, c, d) {
-                    return 0 === a ? b : c * Math.pow(2, 10 * (a / d - 1)) + b
-                }], "ease-out-expo": ["cubic-bezier(0.190, 1, 0.220, 1)", function (a, b, c, d) {
-                    return a === d ? b + c : c * (-Math.pow(2, -10 * a / d) + 1) + b
-                }], "ease-in-out-expo": ["cubic-bezier(1, 0, 0, 1)", function (a, b, c, d) {
-                    return 0 === a ? b : a === d ? b + c : (a /= d / 2) < 1 ? c / 2 * Math.pow(2, 10 * (a - 1)) + b : c / 2 * (-Math.pow(2, -10 * --a) + 2) + b
-                }], "ease-in-circ": ["cubic-bezier(0.600, 0.040, 0.980, 0.335)", function (a, b, c, d) {
-                    return -c * (Math.sqrt(1 - (a /= d) * a) - 1) + b
-                }], "ease-out-circ": ["cubic-bezier(0.075, 0.820, 0.165, 1)", function (a, b, c, d) {
-                    return c * Math.sqrt(1 - (a = a / d - 1) * a) + b
-                }], "ease-in-out-circ": ["cubic-bezier(0.785, 0.135, 0.150, 0.860)", function (a, b, c, d) {
-                    return (a /= d / 2) < 1 ? -c / 2 * (Math.sqrt(1 - a * a) - 1) + b : c / 2 * (Math.sqrt(1 - (a -= 2) * a) + 1) + b
-                }], "ease-in-back": ["cubic-bezier(0.600, -0.280, 0.735, 0.045)", function (a, b, c, d, e) {
-                    return void 0 === e && (e = 1.70158), c * (a /= d) * a * ((e + 1) * a - e) + b
-                }], "ease-out-back": ["cubic-bezier(0.175, 0.885, 0.320, 1.275)", function (a, b, c, d, e) {
-                    return void 0 === e && (e = 1.70158), c * ((a = a / d - 1) * a * ((e + 1) * a + e) + 1) + b
-                }], "ease-in-out-back": ["cubic-bezier(0.680, -0.550, 0.265, 1.550)", function (a, b, c, d, e) {
-                    return void 0 === e && (e = 1.70158), (a /= d / 2) < 1 ? c / 2 * a * a * (((e *= 1.525) + 1) * a - e) + b : c / 2 * ((a -= 2) * a * (((e *= 1.525) + 1) * a + e) + 2) + b
+                return a
+            }("prototype", {}.hasOwnProperty),
+            h = {
+                ease: ["ease", function(t, e, n, i) {
+                    var r = (t /= i) * t,
+                        o = r * t;
+                    return e + n * (-2.75 * o * r + 11 * r * r + -15.5 * o + 8 * r + .25 * t)
+                }],
+                "ease-in": ["ease-in", function(t, e, n, i) {
+                    var r = (t /= i) * t,
+                        o = r * t;
+                    return e + n * (-1 * o * r + 3 * r * r + -3 * o + 2 * r)
+                }],
+                "ease-out": ["ease-out", function(t, e, n, i) {
+                    var r = (t /= i) * t,
+                        o = r * t;
+                    return e + n * (.3 * o * r + -1.6 * r * r + 2.2 * o + -1.8 * r + 1.9 * t)
+                }],
+                "ease-in-out": ["ease-in-out", function(t, e, n, i) {
+                    var r = (t /= i) * t,
+                        o = r * t;
+                    return e + n * (2 * o * r + -5 * r * r + 2 * o + 2 * r)
+                }],
+                linear: ["linear", function(t, e, n, i) {
+                    return n * t / i + e
+                }],
+                "ease-in-quad": ["cubic-bezier(0.550, 0.085, 0.680, 0.530)", function(t, e, n, i) {
+                    return n * (t /= i) * t + e
+                }],
+                "ease-out-quad": ["cubic-bezier(0.250, 0.460, 0.450, 0.940)", function(t, e, n, i) {
+                    return -n * (t /= i) * (t - 2) + e
+                }],
+                "ease-in-out-quad": ["cubic-bezier(0.455, 0.030, 0.515, 0.955)", function(t, e, n, i) {
+                    return (t /= i / 2) < 1 ? n / 2 * t * t + e : -n / 2 * (--t * (t - 2) - 1) + e
+                }],
+                "ease-in-cubic": ["cubic-bezier(0.550, 0.055, 0.675, 0.190)", function(t, e, n, i) {
+                    return n * (t /= i) * t * t + e
+                }],
+                "ease-out-cubic": ["cubic-bezier(0.215, 0.610, 0.355, 1)", function(t, e, n, i) {
+                    return n * ((t = t / i - 1) * t * t + 1) + e
+                }],
+                "ease-in-out-cubic": ["cubic-bezier(0.645, 0.045, 0.355, 1)", function(t, e, n, i) {
+                    return (t /= i / 2) < 1 ? n / 2 * t * t * t + e : n / 2 * ((t -= 2) * t * t + 2) + e
+                }],
+                "ease-in-quart": ["cubic-bezier(0.895, 0.030, 0.685, 0.220)", function(t, e, n, i) {
+                    return n * (t /= i) * t * t * t + e
+                }],
+                "ease-out-quart": ["cubic-bezier(0.165, 0.840, 0.440, 1)", function(t, e, n, i) {
+                    return -n * ((t = t / i - 1) * t * t * t - 1) + e
+                }],
+                "ease-in-out-quart": ["cubic-bezier(0.770, 0, 0.175, 1)", function(t, e, n, i) {
+                    return (t /= i / 2) < 1 ? n / 2 * t * t * t * t + e : -n / 2 * ((t -= 2) * t * t * t -
+                        2) + e
+                }],
+                "ease-in-quint": ["cubic-bezier(0.755, 0.050, 0.855, 0.060)", function(t, e, n, i) {
+                    return n * (t /= i) * t * t * t * t + e
+                }],
+                "ease-out-quint": ["cubic-bezier(0.230, 1, 0.320, 1)", function(t, e, n, i) {
+                    return n * ((t = t / i - 1) * t * t * t * t + 1) + e
+                }],
+                "ease-in-out-quint": ["cubic-bezier(0.860, 0, 0.070, 1)", function(t, e, n, i) {
+                    return (t /= i / 2) < 1 ? n / 2 * t * t * t * t * t + e : n / 2 * ((t -= 2) * t * t * t *
+                        t + 2) + e
+                }],
+                "ease-in-sine": ["cubic-bezier(0.470, 0, 0.745, 0.715)", function(t, e, n, i) {
+                    return -n * Math.cos(t / i * (Math.PI / 2)) + n + e
+                }],
+                "ease-out-sine": ["cubic-bezier(0.390, 0.575, 0.565, 1)", function(t, e, n, i) {
+                    return n * Math.sin(t / i * (Math.PI / 2)) + e
+                }],
+                "ease-in-out-sine": ["cubic-bezier(0.445, 0.050, 0.550, 0.950)", function(t, e, n, i) {
+                    return -n / 2 * (Math.cos(Math.PI * t / i) - 1) + e
+                }],
+                "ease-in-expo": ["cubic-bezier(0.950, 0.050, 0.795, 0.035)", function(t, e, n, i) {
+                    return 0 === t ? e : n * Math.pow(2, 10 * (t / i - 1)) + e
+                }],
+                "ease-out-expo": ["cubic-bezier(0.190, 1, 0.220, 1)", function(t, e, n, i) {
+                    return t === i ? e + n : n * (1 - Math.pow(2, -10 * t / i)) + e
+                }],
+                "ease-in-out-expo": ["cubic-bezier(1, 0, 0, 1)", function(t, e, n, i) {
+                    return 0 === t ? e : t === i ? e + n : (t /= i / 2) < 1 ? n / 2 * Math.pow(2, 10 * (t -
+                        1)) + e : n / 2 * (2 - Math.pow(2, -10 * --t)) + e
+                }],
+                "ease-in-circ": ["cubic-bezier(0.600, 0.040, 0.980, 0.335)", function(t, e, n, i) {
+                    return -n * (Math.sqrt(1 - (t /= i) * t) - 1) + e
+                }],
+                "ease-out-circ": ["cubic-bezier(0.075, 0.820, 0.165, 1)", function(t, e, n, i) {
+                    return n * Math.sqrt(1 - (t = t / i - 1) * t) + e
+                }],
+                "ease-in-out-circ": ["cubic-bezier(0.785, 0.135, 0.150, 0.860)", function(t, e, n, i) {
+                    return (t /= i / 2) < 1 ? -n / 2 * (Math.sqrt(1 - t * t) - 1) + e : n / 2 * (Math.sqrt(
+                        1 - (t -= 2) * t) + 1) + e
+                }],
+                "ease-in-back": ["cubic-bezier(0.600, -0.280, 0.735, 0.045)", function(t, e, n, i, r) {
+                    return void 0 === r && (r = 1.70158), n * (t /= i) * t * ((r + 1) * t - r) + e
+                }],
+                "ease-out-back": ["cubic-bezier(0.175, 0.885, 0.320, 1.275)", function(t, e, n, i, r) {
+                    return void 0 === r && (r = 1.70158), n * ((t = t / i - 1) * t * ((r + 1) * t + r) + 1) +
+                        e
+                }],
+                "ease-in-out-back": ["cubic-bezier(0.680, -0.550, 0.265, 1.550)", function(t, e, n, i, r) {
+                    return void 0 === r && (r = 1.70158), (t /= i / 2) < 1 ? n / 2 * t * t * ((1 + (r *=
+                            1.525)) * t - r) + e : n / 2 * ((t -= 2) * t * ((1 + (r *= 1.525)) * t + r) + 2) +
+                        e
                 }]
-            }, n = {
+            },
+            d = {
                 "ease-in-back": "cubic-bezier(0.600, 0, 0.735, 0.045)",
                 "ease-out-back": "cubic-bezier(0.175, 0.885, 0.320, 1)",
                 "ease-in-out-back": "cubic-bezier(0.680, 0, 0.265, 1)"
-            }, o = document, p = window, q = "bkwld-tram", r = /[\-\.0-9]/g, s = /[A-Z]/, t = "number", u = /^(rgb|#)/,
-            v = /(em|cm|mm|in|pt|pc|px)$/, w = /(em|cm|mm|in|pt|pc|px|%)$/, x = /(deg|rad|turn)$/, y = "unitless",
-            z = /(all|none) 0s ease 0s/, A = /^(width|height)$/, B = " ", C = o.createElement("a"),
-            D = ["Webkit", "Moz", "O", "ms"], E = ["-webkit-", "-moz-", "-o-", "-ms-"], F = function (a) {
-                if (a in C.style)return {dom: a, css: a};
-                var b, c, d = "", e = a.split("-");
-                for (b = 0; b < e.length; b++)d += e[b].charAt(0).toUpperCase() + e[b].slice(1);
-                for (b = 0; b < D.length; b++)if (c = D[b] + d, c in C.style)return {dom: c, css: E[b] + a}
-            }, G = b.support = {
+            },
+            p = document,
+            v = window,
+            m = "bkwld-tram",
+            g = /[\-\.0-9]/g,
+            w = /[A-Z]/,
+            b = "number",
+            y = /^(rgb|#)/,
+            x = /(em|cm|mm|in|pt|pc|px)$/,
+            _ = /(em|cm|mm|in|pt|pc|px|%)$/,
+            k = /(deg|rad|turn)$/,
+            T = "unitless",
+            O = /(all|none) 0s ease 0s/,
+            E = /^(width|height)$/,
+            z = " ",
+            A = p.createElement("a"),
+            j = ["Webkit", "Moz", "O", "ms"],
+            q = ["-webkit-", "-moz-", "-o-", "-ms-"],
+            M = function(t) {
+                if (t in A.style) return {
+                    dom: t,
+                    css: t
+                };
+                var e, n, i = "",
+                    r = t.split("-");
+                for (e = 0; e < r.length; e++) i += r[e].charAt(0).toUpperCase() + r[e].slice(1);
+                for (e = 0; e < j.length; e++)
+                    if ((n = j[e] + i) in A.style) return {
+                        dom: n,
+                        css: q[e] + t
+                    }
+            },
+            $ = e.support = {
                 bind: Function.prototype.bind,
-                transform: F("transform"),
-                transition: F("transition"),
-                backface: F("backface-visibility"),
-                timing: F("transition-timing-function")
+                transform: M("transform"),
+                transition: M("transition"),
+                backface: M("backface-visibility"),
+                timing: M("transition-timing-function")
             };
-        if (G.transition) {
-            var H = G.timing.dom;
-            if (C.style[H] = m["ease-in-back"][0], !C.style[H])for (var I in n)m[I][0] = n[I]
+        if ($.transition) {
+            var D = $.timing.dom;
+            if (A.style[D] = h["ease-in-back"][0], !A.style[D])
+                for (var L in d) h[L][0] = d[L]
         }
-        var J = b.frame = function () {
-            var a = p.requestAnimationFrame || p.webkitRequestAnimationFrame || p.mozRequestAnimationFrame || p.oRequestAnimationFrame || p.msRequestAnimationFrame;
-            return a && G.bind ? a.bind(p) : function (a) {
-                p.setTimeout(a, 16)
-            }
-        }(), K = b.now = function () {
-            var a = p.performance, b = a && (a.now || a.webkitNow || a.msNow || a.mozNow);
-            return b && G.bind ? b.bind(a) : Date.now || function () {
+        var C = e.frame = function() {
+                var t = v.requestAnimationFrame || v.webkitRequestAnimationFrame || v.mozRequestAnimationFrame || v
+                    .oRequestAnimationFrame || v.msRequestAnimationFrame;
+                return t && $.bind ? t.bind(v) : function(t) {
+                    v.setTimeout(t, 16)
+                }
+            }(),
+            R = e.now = function() {
+                var t = v.performance,
+                    e = t && (t.now || t.webkitNow || t.msNow || t.mozNow);
+                return e && $.bind ? e.bind(t) : Date.now || function() {
                     return +new Date
                 }
-        }(), L = l(function (b) {
-            function d(a, b) {
-                var c = k(("" + a).split(B)), d = c[0];
-                b = b || {};
-                var e = Y[d];
-                if (!e)return j("Unsupported property: " + d);
-                if (!b.weak || !this.props[d]) {
-                    var f = e[0], g = this.props[d];
-                    return g || (g = this.props[d] = new f.Bare), g.init(this.$el, c, e, b), g
+            }(),
+            B = f(function(e) {
+                function i(t, e) {
+                    var n = l(("" + t).split(z)),
+                        i = n[0];
+                    e = e || {};
+                    var r = Q[i];
+                    if (!r) return c("Unsupported property: " + i);
+                    if (!e.weak || !this.props[i]) {
+                        var o = r[0],
+                            a = this.props[i];
+                        return a || (a = this.props[i] = new o.Bare), a.init(this.$el, n, r, e), a
+                    }
                 }
+
+                function r(t, e, n) {
+                    if (t) {
+                        var r = typeof t;
+                        if (e || (this.timer && this.timer.destroy(), this.queue = [], this.active = !1),
+                            "number" == r && e) return this.timer = new P({
+                            duration: t,
+                            context: this,
+                            complete: o
+                        }), void(this.active = !0);
+                        if ("string" == r && e) {
+                            switch (t) {
+                                case "hide":
+                                    s.call(this);
+                                    break;
+                                case "stop":
+                                    a.call(this);
+                                    break;
+                                case "redraw":
+                                    f.call(this);
+                                    break;
+                                default:
+                                    i.call(this, t, n && n[1])
+                            }
+                            return o.call(this)
+                        }
+                        if ("function" == r) return void t.call(this, this);
+                        if ("object" == r) {
+                            var c = 0;
+                            d.call(this, t, function(t, e) {
+                                t.span > c && (c = t.span), t.stop(), t.animate(e)
+                            }, function(t) {
+                                "wait" in t && (c = u(t.wait, 0))
+                            }), h.call(this), c > 0 && (this.timer = new P({
+                                duration: c,
+                                context: this
+                            }), this.active = !0, e && (this.timer.complete = o));
+                            var l = this,
+                                p = !1,
+                                v = {};
+                            C(function() {
+                                d.call(l, t, function(t) {
+                                    t.active && (p = !0, v[t.name] = t.nextStyle)
+                                }), p && l.$el.css(v)
+                            })
+                        }
+                    }
+                }
+
+                function o() {
+                    if (this.timer && this.timer.destroy(), this.active = !1, this.queue.length) {
+                        var t = this.queue.shift();
+                        r.call(this, t.options, !0, t.args)
+                    }
+                }
+
+                function a(t) {
+                    this.timer && this.timer.destroy(), this.queue = [], this.active = !1;
+                    var e;
+                    "string" == typeof t ? (e = {}, e[t] = 1) : e = "object" == typeof t && null != t ? t :
+                        this.props, d.call(this, e, p), h.call(this)
+                }
+
+                function s() {
+                    a.call(this), this.el.style.display = "none"
+                }
+
+                function f() {
+                    this.el.offsetHeight
+                }
+
+                function h() {
+                    var t, e, n = [];
+                    this.upstream && n.push(this.upstream);
+                    for (t in this.props)(e = this.props[t]).active && n.push(e.string);
+                    n = n.join(","), this.style !== n && (this.style = n, this.el.style[$.transition.dom] = n)
+                }
+
+                function d(t, e, r) {
+                    var o, a, s, u, c = e !== p,
+                        l = {};
+                    for (o in t) s = t[o], o in J ? (l.transform || (l.transform = {}), l.transform[o] = s) : (
+                        w.test(o) && (o = n(o)), o in Q ? l[o] = s : (u || (u = {}), u[o] = s));
+                    for (o in l) {
+                        if (s = l[o], !(a = this.props[o])) {
+                            if (!c) continue;
+                            a = i.call(this, o)
+                        }
+                        e.call(this, a, s)
+                    }
+                    r && u && r.call(this, u)
+                }
+
+                function p(t) {
+                    t.stop()
+                }
+
+                function v(t, e) {
+                    t.set(e)
+                }
+
+                function g(t) {
+                    this.$el.css(t)
+                }
+
+                function b(t, n) {
+                    e[t] = function() {
+                        return this.children ? y.call(this, n, arguments) : (this.el && n.apply(this,
+                            arguments), this)
+                    }
+                }
+
+                function y(t, e) {
+                    var n, i = this.children.length;
+                    for (n = 0; i > n; n++) t.apply(this.children[n], e);
+                    return this
+                }
+                e.init = function(e) {
+                    if (this.$el = t(e), this.el = this.$el[0], this.props = {}, this.queue = [], this.style =
+                        "", this.active = !1, Z.keepInherited && !Z.fallback) {
+                        var n = W(this.el, "transition");
+                        n && !O.test(n) && (this.upstream = n)
+                    }
+                    $.backface && Z.hideBackface && Y(this.el, $.backface.css, "hidden")
+                }, b("add", i), b("start", r), b("wait", function(t) {
+                    t = u(t, 0), this.active ? this.queue.push({
+                        options: t
+                    }) : (this.timer = new P({
+                        duration: t,
+                        context: this,
+                        complete: o
+                    }), this.active = !0)
+                }), b("then", function(t) {
+                    return this.active ? (this.queue.push({
+                        options: t,
+                        args: arguments
+                    }), void(this.timer.complete = o)) : c(
+                        "No active transition timer. Use start() or wait() before then().")
+                }), b("next", o), b("stop", a), b("set", function(t) {
+                    a.call(this, t), d.call(this, t, v, g)
+                }), b("show", function(t) {
+                    "string" != typeof t && (t = "block"), this.el.style.display = t
+                }), b("hide", s), b("redraw", f), b("destroy", function() {
+                    a.call(this), t.removeData(this.el, m), this.$el = this.el = null
+                })
+            }),
+            S = f(B, function(e) {
+                function n(e, n) {
+                    var i = t.data(e, m) || t.data(e, m, new B.Bare);
+                    return i.el || i.init(e), n ? i.start(n) : i
+                }
+                e.init = function(e, i) {
+                    var r = t(e);
+                    if (!r.length) return this;
+                    if (1 === r.length) return n(r[0], i);
+                    var o = [];
+                    return r.each(function(t, e) {
+                        o.push(n(e, i))
+                    }), this.children = o, this
+                }
+            }),
+            F = f(function(t) {
+                function e() {
+                    var t = this.get();
+                    this.update("auto");
+                    var e = this.get();
+                    return this.update(t), e
+                }
+
+                function n(t, e, n) {
+                    return void 0 !== e && (n = e), t in h ? t : n
+                }
+
+                function i(t) {
+                    var e = /rgba?\((\d+),\s*(\d+),\s*(\d+)/.exec(t);
+                    return (e ? r(e[1], e[2], e[3]) : t).replace(/#(\w)(\w)(\w)$/, "#$1$1$2$2$3$3")
+                }
+                var o = {
+                    duration: 500,
+                    ease: "ease",
+                    delay: 0
+                };
+                t.init = function(t, e, i, r) {
+                    this.$el = t, this.el = t[0];
+                    var a = e[0];
+                    i[2] && (a = i[2]), G[a] && (a = G[a]), this.name = a, this.type = i[1], this.duration =
+                        u(e[1], this.duration, o.duration), this.ease = n(e[2], this.ease, o.ease), this.delay =
+                        u(e[3], this.delay, o.delay), this.span = this.duration + this.delay, this.active = !
+                        1, this.nextStyle = null, this.auto = E.test(this.name), this.unit = r.unit || this
+                        .unit || Z.defaultUnit, this.angle = r.angle || this.angle || Z.defaultAngle, Z.fallback ||
+                        r.fallback ? this.animate = this.fallback : (this.animate = this.transition, this.string =
+                            this.name + z + this.duration + "ms" + ("ease" != this.ease ? z + h[this.ease][
+                                0
+                            ] : "") + (this.delay ? z + this.delay + "ms" : ""))
+                }, t.set = function(t) {
+                    t = this.convert(t, this.type), this.update(t), this.redraw()
+                }, t.transition = function(t) {
+                    this.active = !0, t = this.convert(t, this.type), this.auto && ("auto" == this.el.style[
+                        this.name] && (this.update(this.get()), this.redraw()), "auto" == t && (t =
+                        e.call(this))), this.nextStyle = t
+                }, t.fallback = function(t) {
+                    var n = this.el.style[this.name] || this.convert(this.get(), this.type);
+                    t = this.convert(t, this.type), this.auto && ("auto" == n && (n = this.convert(this.get(),
+                        this.type)), "auto" == t && (t = e.call(this))), this.tween = new X({
+                        from: n,
+                        to: t,
+                        duration: this.duration,
+                        delay: this.delay,
+                        ease: this.ease,
+                        update: this.update,
+                        context: this
+                    })
+                }, t.get = function() {
+                    return W(this.el, this.name)
+                }, t.update = function(t) {
+                    Y(this.el, this.name, t)
+                }, t.stop = function() {
+                    (this.active || this.nextStyle) && (this.active = !1, this.nextStyle = null, Y(this.el,
+                        this.name, this.get()));
+                    var t = this.tween;
+                    t && t.context && t.destroy()
+                }, t.convert = function(t, e) {
+                    if ("auto" == t && this.auto) return t;
+                    var n, r = "number" == typeof t,
+                        o = "string" == typeof t;
+                    switch (e) {
+                        case b:
+                            if (r) return t;
+                            if (o && "" === t.replace(g, "")) return +t;
+                            n = "number(unitless)";
+                            break;
+                        case y:
+                            if (o) {
+                                if ("" === t && this.original) return this.original;
+                                if (e.test(t)) return "#" == t.charAt(0) && 7 == t.length ? t : i(t)
+                            }
+                            n = "hex or rgb string";
+                            break;
+                        case x:
+                            if (r) return t + this.unit;
+                            if (o && e.test(t)) return t;
+                            n = "number(px) or string(unit)";
+                            break;
+                        case _:
+                            if (r) return t + this.unit;
+                            if (o && e.test(t)) return t;
+                            n = "number(px) or string(unit or %)";
+                            break;
+                        case k:
+                            if (r) return t + this.angle;
+                            if (o && e.test(t)) return t;
+                            n = "number(deg) or string(angle)";
+                            break;
+                        case T:
+                            if (r) return t;
+                            if (o && _.test(t)) return t;
+                            n = "number(unitless) or string(unit or %)"
+                    }
+                    return a(n, t), t
+                }, t.redraw = function() {
+                    this.el.offsetHeight
+                }
+            }),
+            N = f(F, function(t, e) {
+                t.init = function() {
+                    e.init.apply(this, arguments), this.original || (this.original = this.convert(this.get(),
+                        y))
+                }
+            }),
+            H = f(F, function(t, e) {
+                t.init = function() {
+                    e.init.apply(this, arguments), this.animate = this.fallback
+                }, t.get = function() {
+                    return this.$el[this.name]()
+                }, t.update = function(t) {
+                    this.$el[this.name](t)
+                }
+            }),
+            I = f(F, function(t, e) {
+                function n(t, e) {
+                    var n, i, r, o, a;
+                    for (n in t) o = J[n], r = o[0], i = o[1] || n, a = this.convert(t[n], r), e.call(this, i,
+                        a, r)
+                }
+                t.init = function() {
+                    e.init.apply(this, arguments), this.current || (this.current = {}, J.perspective && Z.perspective &&
+                        (this.current.perspective = Z.perspective, Y(this.el, this.name, this.style(
+                            this.current)), this.redraw()))
+                }, t.set = function(t) {
+                    n.call(this, t, function(t, e) {
+                        this.current[t] = e
+                    }), Y(this.el, this.name, this.style(this.current)), this.redraw()
+                }, t.transition = function(t) {
+                    var e = this.values(t);
+                    this.tween = new U({
+                        current: this.current,
+                        values: e,
+                        duration: this.duration,
+                        delay: this.delay,
+                        ease: this.ease
+                    });
+                    var n, i = {};
+                    for (n in this.current) i[n] = n in e ? e[n] : this.current[n];
+                    this.active = !0, this.nextStyle = this.style(i)
+                }, t.fallback = function(t) {
+                    var e = this.values(t);
+                    this.tween = new U({
+                        current: this.current,
+                        values: e,
+                        duration: this.duration,
+                        delay: this.delay,
+                        ease: this.ease,
+                        update: this.update,
+                        context: this
+                    })
+                }, t.update = function() {
+                    Y(this.el, this.name, this.style(this.current))
+                }, t.style = function(t) {
+                    var e, n = "";
+                    for (e in t) n += e + "(" + t[e] + ") ";
+                    return n
+                }, t.values = function(t) {
+                    var e, i = {};
+                    return n.call(this, t, function(t, n, r) {
+                        i[t] = n, void 0 === this.current[t] && (e = 0, ~t.indexOf("scale") && (e =
+                            1), this.current[t] = this.convert(e, r))
+                    }), i
+                }
+            }),
+            X = f(function(e) {
+                function n(t) {
+                    1 === d.push(t) && C(a)
+                }
+
+                function a() {
+                    var t, e, n, i = d.length;
+                    if (i)
+                        for (C(a), e = R(), t = i; t--;)(n = d[t]) && n.render(e)
+                }
+
+                function u(e) {
+                    var n, i = t.inArray(e, d);
+                    i >= 0 && (n = d.slice(i + 1), d.length = i, n.length && (d = d.concat(n)))
+                }
+
+                function c(t) {
+                    return Math.round(t * p) / p
+                }
+
+                function l(t, e, n) {
+                    return r(t[0] + n * (e[0] - t[0]), t[1] + n * (e[1] - t[1]), t[2] + n * (e[2] - t[2]))
+                }
+                var f = {
+                    ease: h.ease[1],
+                    from: 0,
+                    to: 1
+                };
+                e.init = function(t) {
+                    this.duration = t.duration || 0, this.delay = t.delay || 0;
+                    var e = t.ease || f.ease;
+                    h[e] && (e = h[e][1]), "function" != typeof e && (e = f.ease), this.ease = e, this.update =
+                        t.update || o, this.complete = t.complete || o, this.context = t.context || this,
+                        this.name = t.name;
+                    var n = t.from,
+                        i = t.to;
+                    void 0 === n && (n = f.from), void 0 === i && (i = f.to), this.unit = t.unit || "",
+                        "number" == typeof n && "number" == typeof i ? (this.begin = n, this.change = i - n) :
+                        this.format(i, n), this.value = this.begin + this.unit, this.start = R(), !1 !== t.autoplay &&
+                        this.play()
+                }, e.play = function() {
+                    this.active || (this.start || (this.start = R()), this.active = !0, n(this))
+                }, e.stop = function() {
+                    this.active && (this.active = !1, u(this))
+                }, e.render = function(t) {
+                    var e, n = t - this.start;
+                    if (this.delay) {
+                        if (n <= this.delay) return;
+                        n -= this.delay
+                    }
+                    if (n < this.duration) {
+                        var i = this.ease(n, 0, 1, this.duration);
+                        return e = this.startRGB ? l(this.startRGB, this.endRGB, i) : c(this.begin + i *
+                            this.change), this.value = e + this.unit, void this.update.call(this.context,
+                            this.value)
+                    }
+                    e = this.endHex || this.begin + this.change, this.value = e + this.unit, this.update.call(
+                        this.context, this.value), this.complete.call(this.context), this.destroy()
+                }, e.format = function(t, e) {
+                    if (e += "", "#" == (t += "").charAt(0)) return this.startRGB = i(e), this.endRGB = i(t),
+                        this.endHex = t, this.begin = 0, void(this.change = 1);
+                    if (!this.unit) {
+                        var n = e.replace(g, "");
+                        n !== t.replace(g, "") && s("tween", e, t), this.unit = n
+                    }
+                    e = parseFloat(e), t = parseFloat(t), this.begin = this.value = e, this.change = t - e
+                }, e.destroy = function() {
+                    this.stop(), this.context = null, this.ease = this.update = this.complete = o
+                };
+                var d = [],
+                    p = 1e3
+            }),
+            P = f(X, function(t) {
+                t.init = function(t) {
+                    this.duration = t.duration || 0, this.complete = t.complete || o, this.context = t.context,
+                        this.play()
+                }, t.render = function(t) {
+                    t - this.start < this.duration || (this.complete.call(this.context), this.destroy())
+                }
+            }),
+            U = f(X, function(t, e) {
+                t.init = function(t) {
+                    this.context = t.context, this.update = t.update, this.tweens = [], this.current = t.current;
+                    var e, n;
+                    for (e in t.values) n = t.values[e], this.current[e] !== n && this.tweens.push(new X({
+                        name: e,
+                        from: this.current[e],
+                        to: n,
+                        duration: t.duration,
+                        delay: t.delay,
+                        ease: t.ease,
+                        autoplay: !1
+                    }));
+                    this.play()
+                }, t.render = function(t) {
+                    var e, n, i = !1;
+                    for (e = this.tweens.length; e--;)(n = this.tweens[e]).context && (n.render(t), this.current[
+                        n.name] = n.value, i = !0);
+                    return i ? void(this.update && this.update.call(this.context)) : this.destroy()
+                }, t.destroy = function() {
+                    if (e.destroy.call(this), this.tweens) {
+                        var t;
+                        for (t = this.tweens.length; t--;) this.tweens[t].destroy();
+                        this.tweens = null, this.current = null
+                    }
+                }
+            }),
+            Z = e.config = {
+                debug: !1,
+                defaultUnit: "px",
+                defaultAngle: "deg",
+                keepInherited: !1,
+                hideBackface: !1,
+                perspective: "",
+                fallback: !$.transition,
+                agentTests: []
+            };
+        e.fallback = function(t) {
+            if (!$.transition) return Z.fallback = !0;
+            Z.agentTests.push("(" + t + ")");
+            var e = new RegExp(Z.agentTests.join("|"), "i");
+            Z.fallback = e.test(navigator.userAgent)
+        }, e.fallback("6.0.[2-5] Safari"), e.tween = function(t) {
+            return new X(t)
+        }, e.delay = function(t, e, n) {
+            return new P({
+                complete: e,
+                duration: t,
+                context: n
+            })
+        }, t.fn.tram = function(t) {
+            return e.call(null, this, t)
+        };
+        var Y = t.style,
+            W = t.css,
+            G = {
+                transform: $.transform && $.transform.css
+            },
+            Q = {
+                color: [N, y],
+                background: [N, y, "background-color"],
+                "outline-color": [N, y],
+                "border-color": [N, y],
+                "border-top-color": [N, y],
+                "border-right-color": [N, y],
+                "border-bottom-color": [N, y],
+                "border-left-color": [N, y],
+                "border-width": [F, x],
+                "border-top-width": [F, x],
+                "border-right-width": [F, x],
+                "border-bottom-width": [F, x],
+                "border-left-width": [F, x],
+                "border-spacing": [F, x],
+                "letter-spacing": [F, x],
+                margin: [F, x],
+                "margin-top": [F, x],
+                "margin-right": [F, x],
+                "margin-bottom": [F, x],
+                "margin-left": [F, x],
+                padding: [F, x],
+                "padding-top": [F, x],
+                "padding-right": [F, x],
+                "padding-bottom": [F, x],
+                "padding-left": [F, x],
+                "outline-width": [F, x],
+                opacity: [F, b],
+                top: [F, _],
+                right: [F, _],
+                bottom: [F, _],
+                left: [F, _],
+                "font-size": [F, _],
+                "text-indent": [F, _],
+                "word-spacing": [F, _],
+                width: [F, _],
+                "min-width": [F, _],
+                "max-width": [F, _],
+                height: [F, _],
+                "min-height": [F, _],
+                "max-height": [F, _],
+                "line-height": [F, T],
+                "scroll-top": [H, b, "scrollTop"],
+                "scroll-left": [H, b, "scrollLeft"]
+            },
+            J = {};
+        $.transform && (Q.transform = [I], J = {
+            x: [_, "translateX"],
+            y: [_, "translateY"],
+            rotate: [k],
+            rotateX: [k],
+            rotateY: [k],
+            scale: [b],
+            scaleX: [b],
+            scaleY: [b],
+            skew: [k],
+            skewX: [k],
+            skewY: [k]
+        }), $.transform && $.backface && (J.z = [_, "translateZ"], J.rotateZ = [k], J.scaleZ = [b], J.perspective = [
+            x
+        ]);
+        var V = /ms/,
+            K = /s|\./;
+        return t.tram = e
+    }(window.jQuery);
+    var e = {},
+        n = t(function(t) {
+            var n = window.$,
+                i = e && n.tram;
+            /*!
+             * Webflow._ (aka) Underscore.js 1.6.0 (custom build)
+             * _.each
+             * _.map
+             * _.find
+             * _.filter
+             * _.any
+             * _.contains
+             * _.delay
+             * _.defer
+             * _.throttle (webflow)
+             * _.debounce
+             * _.keys
+             * _.has
+             * _.now
+             *
+             * http://underscorejs.org
+             * (c) 2009-2013 Jeremy Ashkenas, DocumentCloud and Investigative Reporters & Editors
+             * Underscore may be freely distributed under the MIT license.
+             * @license MIT
+             */
+            t.exports = function() {
+                var t = {};
+                t.VERSION = "1.6.0-Webflow";
+                var e = {},
+                    n = Array.prototype,
+                    r = Object.prototype,
+                    o = Function.prototype,
+                    a = (n.push, n.slice),
+                    s = (n.concat, r.toString, r.hasOwnProperty),
+                    u = n.forEach,
+                    c = n.map,
+                    l = (n.reduce, n.reduceRight, n.filter),
+                    f = (n.every, n.some),
+                    h = n.indexOf,
+                    d = (n.lastIndexOf, Array.isArray, Object.keys),
+                    p = (o.bind, t.each = t.forEach = function(n, i, r) {
+                        if (null == n) return n;
+                        if (u && n.forEach === u) n.forEach(i, r);
+                        else if (n.length === +n.length) {
+                            for (var o = 0, a = n.length; o < a; o++)
+                                if (i.call(r, n[o], o, n) === e) return
+                        } else
+                            for (var s = t.keys(n), o = 0, a = s.length; o < a; o++)
+                                if (i.call(r, n[s[o]], s[o], n) === e) return;
+                        return n
+                    });
+                t.map = t.collect = function(t, e, n) {
+                    var i = [];
+                    return null == t ? i : c && t.map === c ? t.map(e, n) : (p(t, function(t, r, o) {
+                        i.push(e.call(n, t, r, o))
+                    }), i)
+                }, t.find = t.detect = function(t, e, n) {
+                    var i;
+                    return v(t, function(t, r, o) {
+                        if (e.call(n, t, r, o)) return i = t, !0
+                    }), i
+                }, t.filter = t.select = function(t, e, n) {
+                    var i = [];
+                    return null == t ? i : l && t.filter === l ? t.filter(e, n) : (p(t, function(t, r, o) {
+                        e.call(n, t, r, o) && i.push(t)
+                    }), i)
+                };
+                var v = t.some = t.any = function(n, i, r) {
+                    i || (i = t.identity);
+                    var o = !1;
+                    return null == n ? o : f && n.some === f ? n.some(i, r) : (p(n, function(t, n, a) {
+                        if (o || (o = i.call(r, t, n, a))) return e
+                    }), !!o)
+                };
+                t.contains = t.include = function(t, e) {
+                    return null != t && (h && t.indexOf === h ? -1 != t.indexOf(e) : v(t, function(t) {
+                        return t === e
+                    }))
+                }, t.delay = function(t, e) {
+                    var n = a.call(arguments, 2);
+                    return setTimeout(function() {
+                        return t.apply(null, n)
+                    }, e)
+                }, t.defer = function(e) {
+                    return t.delay.apply(t, [e, 1].concat(a.call(arguments, 1)))
+                }, t.throttle = function(t) {
+                    var e, n, r;
+                    return function() {
+                        e || (e = !0, n = arguments, r = this, i.frame(function() {
+                            e = !1, t.apply(r, n)
+                        }))
+                    }
+                }, t.debounce = function(e, n, i) {
+                    var r, o, a, s, u, c = function() {
+                        var l = t.now() - s;
+                        l < n ? r = setTimeout(c, n - l) : (r = null, i || (u = e.apply(a, o), a = o =
+                            null))
+                    };
+                    return function() {
+                        a = this, o = arguments, s = t.now();
+                        var l = i && !r;
+                        return r || (r = setTimeout(c, n)), l && (u = e.apply(a, o), a = o = null), u
+                    }
+                }, t.defaults = function(e) {
+                    if (!t.isObject(e)) return e;
+                    for (var n = 1, i = arguments.length; n < i; n++) {
+                        var r = arguments[n];
+                        for (var o in r) void 0 === e[o] && (e[o] = r[o])
+                    }
+                    return e
+                }, t.keys = function(e) {
+                    if (!t.isObject(e)) return [];
+                    if (d) return d(e);
+                    var n = [];
+                    for (var i in e) t.has(e, i) && n.push(i);
+                    return n
+                }, t.has = function(t, e) {
+                    return s.call(t, e)
+                }, t.isObject = function(t) {
+                    return t === Object(t)
+                }, t.now = Date.now || function() {
+                    return (new Date).getTime()
+                }, t.templateSettings = {
+                    evaluate: /<%([\s\S]+?)%>/g,
+                    interpolate: /<%=([\s\S]+?)%>/g,
+                    escape: /<%-([\s\S]+?)%>/g
+                };
+                var m = /(.)^/,
+                    g = {
+                        "'": "'",
+                        "\\": "\\",
+                        "\r": "r",
+                        "\n": "n",
+                        "\u2028": "u2028",
+                        "\u2029": "u2029"
+                    },
+                    w = /\\|'|\r|\n|\u2028|\u2029/g,
+                    b = function(t) {
+                        return "\\" + g[t]
+                    };
+                return t.template = function(e, n, i) {
+                    !n && i && (n = i), n = t.defaults({}, n, t.templateSettings);
+                    var r = RegExp([(n.escape || m).source, (n.interpolate || m).source, (n.evaluate || m).source]
+                            .join("|") + "|$", "g"),
+                        o = 0,
+                        a = "__p+='";
+                    e.replace(r, function(t, n, i, r, s) {
+                            return a += e.slice(o, s).replace(w, b), o = s + t.length, n ? a +=
+                                "'+\n((__t=(" + n + "))==null?'':_.escape(__t))+\n'" : i ? a +=
+                                "'+\n((__t=(" + i + "))==null?'':__t)+\n'" : r && (a += "';\n" + r +
+                                    "\n__p+='"), t
+                        }), a += "';\n", n.variable || (a = "with(obj||{}){\n" + a + "}\n"), a =
+                        "var __t,__p='',__j=Array.prototype.join,print=function(){__p+=__j.call(arguments,'');};\n" +
+                        a + "return __p;\n";
+                    try {
+                        var s = new Function(n.variable || "obj", "_", a)
+                    } catch (t) {
+                        throw t.source = a, t
+                    }
+                    var u = function(e) {
+                            return s.call(this, e, t)
+                        },
+                        c = n.variable || "obj";
+                    return u.source = "function(" + c + "){\n" + a + "}", u
+                }, t
+            }()
+        }),
+        i = n && "object" == typeof n && "default" in n ? n.default : n,
+        r = t(function(t) {
+            function n(t) {
+                f.env() && (w(t.design) && m.on("__wf_design", t.design), w(t.preview) && m.on("__wf_preview", t.preview)),
+                    w(t.destroy) && m.on("__wf_destroy", t.destroy), t.ready && w(t.ready) && r(t)
             }
 
-            function e(a, b, c) {
-                if (a) {
-                    var e = typeof a;
-                    if (b || (this.timer && this.timer.destroy(), this.queue = [], this.active = !1), "number" == e && b)return this.timer = new S({
-                        duration: a,
-                        context: this,
-                        complete: h
-                    }), void(this.active = !0);
-                    if ("string" == e && b) {
-                        switch (a) {
-                            case"hide":
-                                o.call(this);
-                                break;
-                            case"stop":
-                                l.call(this);
-                                break;
-                            case"redraw":
-                                p.call(this);
-                                break;
-                            default:
-                                d.call(this, a, c && c[1])
-                        }
-                        return h.call(this)
+            function r(t) {
+                x ? t.ready() : b.contains(d, t.ready) || d.push(t.ready)
+            }
+
+            function o(t) {
+                w(t.design) && m.off("__wf_design", t.design), w(t.preview) && m.off("__wf_preview", t.preview), w(
+                    t.destroy) && m.off("__wf_destroy", t.destroy), t.ready && w(t.ready) && a(t)
+            }
+
+            function a(t) {
+                d = b.filter(d, function(e) {
+                    return e !== t.ready
+                })
+            }
+
+            function s(t, e) {
+                var n = [],
+                    i = {};
+                return i.up = b.throttle(function(t) {
+                    b.each(n, function(e) {
+                        e(t)
+                    })
+                }), t && e && t.on(e, i.up), i.on = function(t) {
+                    "function" == typeof t && (b.contains(n, t) || n.push(t))
+                }, i.off = function(t) {
+                    n = arguments.length ? b.filter(n, function(e) {
+                        return e !== t
+                    }) : []
+                }, i
+            }
+
+            function u(t) {
+                w(t) && t()
+            }
+
+            function c() {
+                _ = !1, b.each(h, n)
+            }
+
+            function l() {
+                q && (q.reject(), m.off("load", q.resolve)), q = new v.Deferred, m.on("load", q.resolve)
+            }
+            var f = {},
+                h = {},
+                d = [],
+                p = window.Webflow || [],
+                v = window.jQuery,
+                m = v(window),
+                g = v(document),
+                w = v.isFunction,
+                b = f._ = i,
+                y = e && v.tram,
+                x = !1,
+                _ = !1;
+            y.config.hideBackface = !1, y.config.keepInherited = !0, f.define = function(t, e, i) {
+                h[t] && o(h[t]);
+                var r = h[t] = e(v, b, i) || {};
+                return n(r), r
+            }, f.require = function(t) {
+                return h[t]
+            }, f.push = function(t) {
+                x ? w(t) && t() : p.push(t)
+            }, f.env = function(t) {
+                var e = window.__wf_design,
+                    n = void 0 !== e;
+                return t ? "design" === t ? n && e : "preview" === t ? n && !e : "slug" === t ? n && window.__wf_slug :
+                    "editor" === t ? window.WebflowEditor : "test" === t ? window.__wf_test : "frame" === t ?
+                    window !== window.top : void 0 : n
+            };
+            var k = navigator.userAgent.toLowerCase(),
+                T = navigator.appVersion.toLowerCase(),
+                O = f.env.touch = "ontouchstart" in window || window.DocumentTouch && document instanceof window.DocumentTouch,
+                E = f.env.chrome = /chrome/.test(k) && /Google/.test(navigator.vendor) && parseInt(T.match(
+                    /chrome\/(\d+)\./)[1], 10),
+                z = f.env.ios = /(ipod|iphone|ipad)/.test(k);
+            f.env.safari = /safari/.test(k) && !E && !z;
+            var A;
+            O && g.on("touchstart mousedown", function(t) {
+                A = t.target
+            }), f.validClick = O ? function(t) {
+                return t === A || v.contains(t, A)
+            } : function() {
+                return !0
+            };
+            var j = "resize.webflow orientationchange.webflow load.webflow";
+            f.resize = s(m, j), f.scroll = s(m,
+                    "scroll.webflow resize.webflow orientationchange.webflow load.webflow"), f.redraw = s(), f.location =
+                function(t) {
+                    window.location = t
+                }, f.env() && (f.location = function(t) {
+                    window.dispatchEvent(new CustomEvent("__wf_location", {
+                        detail: t
+                    }))
+                }), f.ready = function() {
+                    x = !0, _ ? c() : b.each(d, u), b.each(p, u), f.resize.up()
+                };
+            var q;
+            f.load = function(t) {
+                q.then(t)
+            }, f.destroy = function(t) {
+                t = t || {}, _ = !0, m.triggerHandler("__wf_destroy"), null != t.domready && (x = t.domready),
+                    b.each(h, o), f.resize.off(), f.scroll.off(), f.redraw.off(), d = [], p = [], "pending" ===
+                    q.state() && l()
+            }, v(f.ready), l(), t.exports = window.Webflow = f
+        }),
+        o = r && "object" == typeof r && "default" in r ? r.default : r,
+        a = (t(function(t) {
+            var e = o;
+            e.define("brand", t.exports = function(t) {
+                function n() {
+                    var e = t('<a class="w-webflow-badge"></a>').attr("href",
+                            "https://webflow.com?utm_campaign=brandjs"),
+                        n = t("<img>").attr("src",
+                            "https://d1otoma47x30pg.cloudfront.net/img/webflow-badge-icon.60efbf6ec9.svg"
+                        ).css({
+                            marginRight: "8px",
+                            width: "16px"
+                        }),
+                        i = t("<img>").attr("src",
+                            "https://d1otoma47x30pg.cloudfront.net/img/webflow-badge-text.6faa6a38cd.svg"
+                        );
+                    return e.append(n, i), e[0]
+                }
+
+                function i() {
+                    var t = s.children(u),
+                        n = t.length && t.get(0) === r,
+                        i = e.env("editor");
+                    n ? i && t.remove() : (t.length && t.remove(), i || s.append(r))
+                }
+                var r, o = {},
+                    a = t("html"),
+                    s = t("body"),
+                    u = ".w-webflow-badge",
+                    c = window.location,
+                    l = /PhantomJS/i.test(navigator.userAgent);
+                return o.ready = function() {
+                    var t = a.attr("data-wf-status"),
+                        e = a.attr("data-wf-domain") || "";
+                    /\.webflow\.io$/i.test(e) && c.hostname !== e && (t = !0), t && !l && (r = r ||
+                        n(), i(), setTimeout(i, 500))
+                }, o
+            })
+        }), t(function(t) {
+            var e = window.jQuery,
+                n = {},
+                i = [],
+                r = {
+                    reset: function(t, e) {
+                        e.__wf_intro = null
+                    },
+                    intro: function(t, i) {
+                        i.__wf_intro || (i.__wf_intro = !0, e(i).triggerHandler(n.types.INTRO))
+                    },
+                    outro: function(t, i) {
+                        i.__wf_intro && (i.__wf_intro = null, e(i).triggerHandler(n.types.OUTRO))
                     }
-                    if ("function" == e)return void a.call(this, this);
-                    if ("object" == e) {
-                        var f = 0;
-                        u.call(this, a, function (a, b) {
-                            a.span > f && (f = a.span), a.stop(), a.animate(b)
-                        }, function (a) {
-                            "wait" in a && (f = i(a.wait, 0))
-                        }), t.call(this), f > 0 && (this.timer = new S({
-                            duration: f,
-                            context: this
-                        }), this.active = !0, b && (this.timer.complete = h));
-                        var g = this, j = !1, k = {};
-                        J(function () {
-                            u.call(g, a, function (a) {
-                                a.active && (j = !0, k[a.name] = a.nextStyle)
-                            }), j && g.$el.css(k)
+                };
+            n.triggers = {}, n.types = {
+                INTRO: "w-ix-intro.w-ix",
+                OUTRO: "w-ix-outro.w-ix"
+            }, n.init = function() {
+                for (var t = i.length, o = 0; o < t; o++) {
+                    var a = i[o];
+                    a[0](0, a[1])
+                }
+                i = [], e.extend(n.triggers, r)
+            }, n.async = function() {
+                for (var t in r) {
+                    var e = r[t];
+                    r.hasOwnProperty(t) && (n.triggers[t] = function(t, n) {
+                        i.push([e, n])
+                    })
+                }
+            }, n.async(), t.exports = n
+        })),
+        s = a && "object" == typeof a && "default" in a ? a.default : a;
+    t(function(t) {
+            var e = o,
+                n = s;
+            e.define("dropdown", t.exports = function(t, i) {
+                function r() {
+                    b = _ && e.env("design"), (w = x.find(T)).each(o)
+                }
+
+                function o(e, n) {
+                    var i = t(n),
+                        r = t.data(n, T);
+                    r || (r = t.data(n, T, {
+                            open: !1,
+                            el: i,
+                            config: {}
+                        })), r.list = i.children(".w-dropdown-list"), r.toggle = i.children(
+                            ".w-dropdown-toggle"), r.links = r.list.children(".w-dropdown-link"), r.outside =
+                        d(r), r.complete = p(r), r.leave = m(r), r.moveOutside = g(r), i.off(T), r.toggle.off(
+                            T), a(r), r.nav && r.nav.off(T), r.nav = i.closest(".w-nav"), r.nav.on(E, s(r)),
+                        b ? i.on("setting" + T, s(r)) : (r.toggle.on("tap" + T, u(r)), r.config.hover && r.toggle
+                            .on("mouseenter" + T, v(r)), i.on(E, s(r)), _ && l(r))
+                }
+
+                function a(t) {
+                    var e = Number(t.el.css("z-index"));
+                    t.manageZ = e === A || e === A + 1, t.config = {
+                        hover: Boolean(t.el.attr("data-hover")) && !k,
+                        delay: Number(t.el.attr("data-delay")) || 0
+                    }
+                }
+
+                function s(t) {
+                    return function(e, n) {
+                        return n = n || {}, "w-close" === e.type ? l(t) : "setting" === e.type ? (a(t), !
+                            0 === n.open && c(t, !0), void(!1 === n.open && l(t, !0))) : void 0
+                    }
+                }
+
+                function u(t) {
+                    return i.debounce(function() {
+                        t.open ? l(t) : c(t)
+                    })
+                }
+
+                function c(t) {
+                    if (!t.open) {
+                        h(t), t.open = !0, t.list.addClass(O), t.toggle.addClass(O), z.intro(0, t.el[0]), e
+                            .redraw.up(), t.manageZ && t.el.css("z-index", A + 1);
+                        var n = e.env("editor");
+                        b || x.on("mouseup" + T, t.outside), t.hovering && !n && t.el.on("mouseleave" + T,
+                            t.leave), t.hovering && n && x.on("mousemove" + T, t.moveOutside), window.clearTimeout(
+                            t.delayId)
+                    }
+                }
+
+                function l(t, e) {
+                    if (t.open && (!t.config.hover || !t.hovering)) {
+                        t.open = !1;
+                        var n = t.config;
+                        if (z.outro(0, t.el[0]), x.off("mouseup" + T, t.outside), t.el.off("mouseleave" + T,
+                                t.leave), x.off("mousemove" + T, t.moveOutside), window.clearTimeout(t.delayId), !
+                            n.delay || e) return t.complete();
+                        t.delayId = window.setTimeout(t.complete, n.delay)
+                    }
+                }
+
+                function f() {
+                    x.find(T).each(function(e, n) {
+                        t(n).triggerHandler(E)
+                    })
+                }
+
+                function h(e) {
+                    var n = e.el[0];
+                    w.each(function(e, i) {
+                        var r = t(i);
+                        r.is(n) || r.has(n).length || r.triggerHandler(E)
+                    })
+                }
+
+                function d(n) {
+                    return n.outside && x.off("mouseup" + T, n.outside), i.debounce(function(i) {
+                        if (n.open) {
+                            var r = t(i.target);
+                            if (!r.closest(".w-dropdown-toggle").length) {
+                                var o = -1 === t.inArray(n.el[0], r.parents(T)),
+                                    a = e.env("editor");
+                                if (o) {
+                                    if (a) {
+                                        var s = 1 === r.parents().length && 1 === r.parents("svg").length,
+                                            u = r.parents(".w-editor-bem-EditorHoverControls").length;
+                                        if (s || u) return
+                                    }
+                                    l(n)
+                                }
+                            }
+                        }
+                    })
+                }
+
+                function p(t) {
+                    return function() {
+                        t.list.removeClass(O), t.toggle.removeClass(O), t.manageZ && t.el.css("z-index",
+                            "")
+                    }
+                }
+
+                function v(t) {
+                    return function() {
+                        t.hovering = !0, c(t)
+                    }
+                }
+
+                function m(t) {
+                    return function() {
+                        t.hovering = !1, l(t)
+                    }
+                }
+
+                function g(e) {
+                    return i.debounce(function(n) {
+                        if (e.open) {
+                            var i = t(n.target);
+                            if (-1 === t.inArray(e.el[0], i.parents(T))) {
+                                var r = i.parents(".w-editor-bem-EditorHoverControls").length,
+                                    o = i.parents(".w-editor-bem-RTToolbar").length,
+                                    a = t(".w-editor-bem-EditorOverlay"),
+                                    s = a.find(".w-editor-edit-outline").length || a.find(
+                                        ".w-editor-bem-RTToolbar").length;
+                                if (r || o || s) return;
+                                e.hovering = !1, l(e)
+                            }
+                        }
+                    })
+                }
+                var w, b, y = {},
+                    x = t(document),
+                    _ = e.env(),
+                    k = e.env.touch,
+                    T = ".w-dropdown",
+                    O = "w--open",
+                    E = "w-close" + T,
+                    z = n.triggers,
+                    A = 900,
+                    j = !1;
+                return y.ready = r, y.design = function() {
+                    j && f(), j = !1, r()
+                }, y.preview = function() {
+                    j = !0, r()
+                }, y
+            })
+        }), t(function(t) {
+            /*!
+             * jQuery-ajaxTransport-XDomainRequest - v1.0.3
+             * 2014-12-16 WEBFLOW - Removed UMD wrapper
+             * https://github.com/MoonScript/jQuery-ajaxTransport-XDomainRequest
+             * Copyright (c) 2014 Jason Moon (@JSONMOON)
+             * @license MIT (/blob/master/LICENSE.txt)
+             */
+            t.exports = function(t) {
+                if (!t.support.cors && t.ajaxTransport && window.XDomainRequest) {
+                    var e = /^https?:\/\//i,
+                        n = /^get|post$/i,
+                        i = new RegExp("^" + location.protocol, "i");
+                    t.ajaxTransport("* text html xml json", function(r, o, a) {
+                        if (r.crossDomain && r.async && n.test(r.type) && e.test(r.url) && i.test(r.url)) {
+                            var s = null;
+                            return {
+                                send: function(e, n) {
+                                    var i = "",
+                                        a = (o.dataType || "").toLowerCase();
+                                    s = new XDomainRequest, /^\d+$/.test(o.timeout) && (s.timeout =
+                                        o.timeout), s.ontimeout = function() {
+                                        n(500, "timeout")
+                                    }, s.onload = function() {
+                                        var e = "Content-Length: " + s.responseText.length +
+                                            "\r\nContent-Type: " + s.contentType,
+                                            i = {
+                                                code: 200,
+                                                message: "success"
+                                            },
+                                            r = {
+                                                text: s.responseText
+                                            };
+                                        try {
+                                            if ("html" === a || /text\/html/i.test(s.contentType))
+                                                r.html = s.responseText;
+                                            else if ("json" === a || "text" !== a && /\/json/i.test(
+                                                    s.contentType)) try {
+                                                r.json = t.parseJSON(s.responseText)
+                                            } catch (t) {
+                                                i.code = 500, i.message = "parseerror"
+                                            } else if ("xml" === a || "text" !== a &&
+                                                /\/xml/i.test(s.contentType)) {
+                                                var o = new ActiveXObject(
+                                                    "Microsoft.XMLDOM");
+                                                o.async = !1;
+                                                try {
+                                                    o.loadXML(s.responseText)
+                                                } catch (t) {
+                                                    o = void 0
+                                                }
+                                                if (!o || !o.documentElement || o.getElementsByTagName(
+                                                        "parsererror").length) throw i.code =
+                                                    500, i.message = "parseerror",
+                                                    "Invalid XML: " + s.responseText;
+                                                r.xml = o
+                                            }
+                                        } catch (t) {
+                                            throw t
+                                        } finally {
+                                            n(i.code, i.message, r, e)
+                                        }
+                                    }, s.onprogress = function() {}, s.onerror = function() {
+                                        n(500, "error", {
+                                            text: s.responseText
+                                        })
+                                    }, o.data && (i = "string" === t.type(o.data) ? o.data : t.param(
+                                        o.data)), s.open(r.type, r.url), s.send(i)
+                                },
+                                abort: function() {
+                                    s && s.abort()
+                                }
+                            }
+                        }
+                    })
+                }
+            }(window.jQuery)
+        }),
+        t(function(t) {
+            var e = o,
+                n = s;
+            e.define("ix", t.exports = function(t, i) {
+                function r(t) {
+                    t && (E = {}, i.each(t, function(t) {
+                        E[t.slug] = t.value
+                    }), o())
+                }
+
+                function o() {
+                    var i = t("[data-ix]");
+                    i.length && (i.each(u), i.each(a), z.length && (e.scroll.on(c), setTimeout(c, 1)), A.length &&
+                        e.load(l), j.length && setTimeout(f, q), n.init(), e.redraw.up())
+                }
+
+                function a(r, o) {
+                    var a = t(o),
+                        u = a.attr("data-ix"),
+                        c = E[u];
+                    if (c) {
+                        var l = c.triggers;
+                        l && (g.style(a, c.style), i.each(l, function(t) {
+                            function i() {
+                                h(t, a, {
+                                    group: "A"
+                                })
+                            }
+
+                            function r() {
+                                h(t, a, {
+                                    group: "B"
+                                })
+                            }
+                            var o = {},
+                                u = t.type,
+                                c = t.stepsB && t.stepsB.length;
+                            if ("load" !== u) {
+                                if ("click" === u) return a.on("click" + b, function(n) {
+                                    e.validClick(n.currentTarget) && ("#" === a.attr(
+                                        "href") && n.preventDefault(), h(t,
+                                        a, {
+                                            group: o.clicked ? "B" : "A"
+                                        }), c && (o.clicked = !o.clicked))
+                                }), void(O = O.add(a));
+                                if ("hover" === u) return a.on("mouseenter" + b, i), a.on(
+                                    "mouseleave" + b, r), void(O = O.add(a));
+                                if ("scroll" !== u) {
+                                    var l = M[u];
+                                    if (l) {
+                                        var f = a.closest(l);
+                                        return f.on(n.types.INTRO, i).on(n.types.OUTRO, r),
+                                            void(O = O.add(f))
+                                    }
+                                } else z.push({
+                                    el: a,
+                                    trigger: t,
+                                    state: {
+                                        active: !1
+                                    },
+                                    offsetTop: s(t.offsetTop),
+                                    offsetBot: s(t.offsetBot)
+                                })
+                            } else t.preload && !_ ? A.push(i) : j.push(i)
+                        }))
+                    }
+                }
+
+                function s(t) {
+                    if (!t) return 0;
+                    t = String(t);
+                    var e = parseInt(t, 10);
+                    return e !== e ? 0 : (t.indexOf("%") > 0 && (e /= 100) >= 1 && (e = .999), e)
+                }
+
+                function u(e, n) {
+                    t(n).off(b)
+                }
+
+                function c() {
+                    for (var t = w.scrollTop(), e = w.height(), n = z.length, i = 0; i < n; i++) {
+                        var r = z[i],
+                            o = r.el,
+                            a = r.trigger,
+                            s = a.stepsB && a.stepsB.length,
+                            u = r.state,
+                            c = o.offset().top,
+                            l = o.outerHeight(),
+                            f = r.offsetTop,
+                            d = r.offsetBot;
+                        f < 1 && f > 0 && (f *= e), d < 1 && d > 0 && (d *= e);
+                        var p = c + l - f >= t && c + d <= t + e;
+                        p !== u.active && ((!1 !== p || s) && (u.active = p, h(a, o, {
+                            group: p ? "A" : "B"
+                        })))
+                    }
+                }
+
+                function l() {
+                    for (var t = A.length, e = 0; e < t; e++) A[e]()
+                }
+
+                function f() {
+                    for (var t = j.length, e = 0; e < t; e++) j[e]()
+                }
+
+                function h(e, n, i, r) {
+                    function o() {
+                        if (c) return h(e, n, i, !0);
+                        "auto" === m.width && p.set({
+                            width: "auto"
+                        }), "auto" === m.height && p.set({
+                            height: "auto"
+                        }), a && a()
+                    }
+                    var a = (i = i || {}).done,
+                        s = e.preserve3d;
+                    if (!v || i.force) {
+                        var u = i.group || "A",
+                            c = e["loop" + u],
+                            l = e["steps" + u];
+                        if (l && l.length) {
+                            if (l.length < 2 && (c = !1), !r) {
+                                var f = e.selector;
+                                f && (n = e.descend ? n.find(f) : e.siblings ? n.siblings(f) : t(f), _ && n
+                                        .attr("data-ix-affect", 1)), k && n.addClass("w-ix-emptyfix"), s &&
+                                    n.css("transform-style", "preserve-3d")
+                            }
+                            for (var p = y(n), m = {
+                                    omit3d: !s
+                                }, g = 0; g < l.length; g++) d(p, l[g], m);
+                            m.start ? p.then(o) : o()
+                        }
+                    }
+                }
+
+                function d(t, n, i) {
+                    var r = "add",
+                        o = "start";
+                    i.start && (r = o = "then");
+                    var a = n.transition;
+                    if (a) {
+                        a = a.split(",");
+                        for (var s = 0; s < a.length; s++) {
+                            var u = a[s];
+                            t[r](u)
+                        }
+                    }
+                    var c = p(n, i) || {};
+                    if (null != c.width && (i.width = c.width), null != c.height && (i.height = c.height),
+                        null == a) {
+                        i.start ? t.then(function() {
+                            var n = this.queue;
+                            this.set(c), c.display && (t.redraw(), e.redraw.up()), this.queue = n,
+                                this.next()
+                        }) : (t.set(c), c.display && (t.redraw(), e.redraw.up()));
+                        var l = c.wait;
+                        null != l && (t.wait(l), i.start = !0)
+                    } else {
+                        if (c.display) {
+                            var f = c.display;
+                            delete c.display, i.start ? t.then(function() {
+                                var t = this.queue;
+                                this.set({
+                                    display: f
+                                }).redraw(), e.redraw.up(), this.queue = t, this.next()
+                            }) : (t.set({
+                                display: f
+                            }).redraw(), e.redraw.up())
+                        }
+                        t[o](c), i.start = !0
+                    }
+                }
+
+                function p(t, e) {
+                    var n = e && e.omit3d,
+                        i = {},
+                        r = !1;
+                    for (var o in t) "transition" !== o && "keysort" !== o && (!n || "z" !== o && "rotateX" !==
+                        o && "rotateY" !== o && "scaleZ" !== o) && (i[o] = t[o], r = !0);
+                    return r ? i : null
+                }
+                var v, m, g = {},
+                    w = t(window),
+                    b = ".w-ix",
+                    y = t.tram,
+                    x = e.env,
+                    _ = x(),
+                    k = x.chrome && x.chrome < 35,
+                    T = "none 0s ease 0s",
+                    O = t(),
+                    E = {},
+                    z = [],
+                    A = [],
+                    j = [],
+                    q = 1,
+                    M = {
+                        tabs: ".w-tab-link, .w-tab-pane",
+                        dropdown: ".w-dropdown",
+                        slider: ".w-slide",
+                        navbar: ".w-nav"
+                    };
+                return g.init = function(t) {
+                    setTimeout(function() {
+                        r(t)
+                    }, 1)
+                }, g.preview = function() {
+                    v = !1, q = 100, setTimeout(function() {
+                        r(window.__wf_ix)
+                    }, 1)
+                }, g.design = function() {
+                    v = !0, g.destroy()
+                }, g.destroy = function() {
+                    m = !0, O.each(u), e.scroll.off(c), n.async(), z = [], A = [], j = []
+                }, g.ready = function() {
+                    if (_) return x("design") ? g.design() : g.preview();
+                    E && m && (m = !1, o())
+                }, g.run = h, g.style = _ ? function(e, n) {
+                    var i = y(e);
+                    if (!t.isEmptyObject(n)) {
+                        e.css("transition", "");
+                        var r = e.css("transition");
+                        r === T && (r = i.upstream = null), i.upstream = T, i.set(p(n)), i.upstream = r
+                    }
+                } : function(t, e) {
+                    y(t).set(p(e))
+                }, g
+            })
+        }), t(function(t) {
+            var e = o;
+            e.define("links", t.exports = function(t, n) {
+                function i(e) {
+                    var n = a && e.getAttribute("href-disabled") || e.getAttribute("href");
+                    if (d.href = n, !(n.indexOf(":") >= 0)) {
+                        var i = t(e);
+                        if (0 === n.indexOf("#") && v.test(n)) {
+                            var r = t(n);
+                            r.length && s.push({
+                                link: i,
+                                sec: r,
+                                active: !1
+                            })
+                        } else if ("#" !== n) {
+                            var c = d.href === h.href || n === u || m.test(n) && g.test(u);
+                            o(i, p, c)
+                        }
+                    }
+                }
+
+                function r() {
+                    var t = l.scrollTop(),
+                        e = l.height();
+                    n.each(s, function(n) {
+                        var i = n.link,
+                            r = n.sec,
+                            a = r.offset().top,
+                            s = r.outerHeight(),
+                            u = .5 * e,
+                            c = r.is(":visible") && a + s - u >= t && a + u <= t + e;
+                        n.active !== c && (n.active = c, o(i, p, c))
+                    })
+                }
+
+                function o(t, e, n) {
+                    var i = t.hasClass(e);
+                    n && i || (n || i) && (n ? t.addClass(e) : t.removeClass(e))
+                }
+                var a, s, u, c = {},
+                    l = t(window),
+                    f = e.env(),
+                    h = window.location,
+                    d = document.createElement("a"),
+                    p = "w--current",
+                    v = /^#[\w:.-]+$/,
+                    m = /index\.(html|php)$/,
+                    g = /\/$/;
+                return c.ready = c.design = c.preview = function() {
+                    a = f && e.env("design"), u = e.env("slug") || h.pathname || "", e.scroll.off(r), s = [];
+                    for (var t = document.links, n = 0; n < t.length; ++n) i(t[n]);
+                    s.length && (e.scroll.on(r), r())
+                }, c
+            })
+        }), t(function(t) {
+            var e = o,
+                n = s;
+            e.define("navbar", t.exports = function(t, i) {
+                function r() {
+                    e.resize.off(a)
+                }
+
+                function o() {
+                    e.resize.on(a)
+                }
+
+                function a() {
+                    k.each(g)
+                }
+
+                function s(e, n) {
+                    var i = t(n),
+                        r = t.data(n, M);
+                    r || (r = t.data(n, M, {
+                            open: !1,
+                            el: i,
+                            config: {}
+                        })), r.menu = i.find(".w-nav-menu"), r.links = r.menu.find(".w-nav-link"), r.dropdowns =
+                        r.menu.find(".w-dropdown"), r.button = i.find(".w-nav-button"), r.container = i.find(
+                            ".w-container"), r.outside = m(r), r.el.off(M), r.button.off(M), r.menu.off(M),
+                        f(r), T ? (c(r), r.el.on("setting" + M, h(r))) : (l(r), r.button.on("tap" + M, p(r)),
+                            r.menu.on("click" + M, "a", v(r))), g(e, n)
+                }
+
+                function u(e, n) {
+                    var i = t.data(n, M);
+                    i && (c(i), t.removeData(n, M))
+                }
+
+                function c(t) {
+                    t.overlay && (x(t, !0), t.overlay.remove(), t.overlay = null)
+                }
+
+                function l(e) {
+                    e.overlay || (e.overlay = t(q).appendTo(e.el), e.parent = e.menu.parent(), x(e, !0))
+                }
+
+                function f(t) {
+                    var e = {},
+                        n = t.config || {},
+                        r = e.animation = t.el.attr("data-animation") || "default";
+                    e.animOver = /^over/.test(r), e.animDirect = /left$/.test(r) ? -1 : 1, n.animation !==
+                        r && t.open && i.defer(d, t), e.easing = t.el.attr("data-easing") || "ease", e.easing2 =
+                        t.el.attr("data-easing2") || "ease";
+                    var o = t.el.attr("data-duration");
+                    e.duration = null != o ? Number(o) : 400, e.docHeight = t.el.attr("data-doc-height"), t
+                        .config = e
+                }
+
+                function h(t) {
+                    return function(e, n) {
+                        n = n || {};
+                        var r = z.width();
+                        f(t), !0 === n.open && b(t, !0), !1 === n.open && x(t, !0), t.open && i.defer(
+                            function() {
+                                r !== z.width() && d(t)
+                            })
+                    }
+                }
+
+                function d(t) {
+                    t.open && (x(t, !0), b(t, !0))
+                }
+
+                function p(t) {
+                    return i.debounce(function() {
+                        t.open ? x(t) : b(t)
+                    })
+                }
+
+                function v(n) {
+                    return function(i) {
+                        var r = t(this).attr("href");
+                        e.validClick(i.currentTarget) ? r && 0 === r.indexOf("#") && n.open && x(n) : i
+                            .preventDefault()
+                    }
+                }
+
+                function m(e) {
+                    return e.outside && A.off("tap" + M, e.outside), i.debounce(function(n) {
+                        if (e.open) {
+                            var i = t(n.target).closest(".w-nav-menu");
+                            e.menu.is(i) || x(e)
+                        }
+                    })
+                }
+
+                function g(e, n) {
+                    var i = t.data(n, M),
+                        r = i.collapsed = "none" !== i.button.css("display");
+                    if (!i.open || r || T || x(i, !0), i.container.length) {
+                        var o = w(i);
+                        i.links.each(o), i.dropdowns.each(o)
+                    }
+                    i.open && y(i)
+                }
+
+                function w(e) {
+                    var n = e.container.css(B);
+                    return "none" === n && (n = ""),
+                        function(e, i) {
+                            (i = t(i)).css(B, ""), "none" === i.css(B) && i.css(B, n)
+                        }
+                }
+
+                function b(t, n) {
+                    if (!t.open) {
+                        t.open = !0, t.menu.addClass(D), t.links.addClass(L), t.button.addClass($);
+                        var i = t.config;
+                        "none" !== i.animation && E.support.transform || (n = !0);
+                        var r = y(t),
+                            o = t.menu.outerHeight(!0),
+                            a = t.menu.outerWidth(!0),
+                            s = t.el.height(),
+                            u = t.el[0];
+                        if (g(0, u), C.intro(0, u), e.redraw.up(), T || A.on("tap" + M, t.outside), !n) {
+                            var c = "transform " + i.duration + "ms " + i.easing;
+                            if (t.overlay && (R = t.menu.prev(), t.overlay.show().append(t.menu)), i.animOver)
+                                return E(t.menu).add(c).set({
+                                    x: i.animDirect * a,
+                                    height: r
+                                }).start({
+                                    x: 0
+                                }), void(t.overlay && t.overlay.width(a));
+                            var l = s + o;
+                            E(t.menu).add(c).set({
+                                y: -l
+                            }).start({
+                                y: 0
+                            })
+                        }
+                    }
+                }
+
+                function y(t) {
+                    var e = t.config,
+                        n = e.docHeight ? A.height() : _.height();
+                    return e.animOver ? t.menu.height(n) : "fixed" !== t.el.css("position") && (n -= t.el.height()),
+                        t.overlay && t.overlay.height(n), n
+                }
+
+                function x(t, e) {
+                    function n() {
+                        t.menu.height(""), E(t.menu).set({
+                                x: 0,
+                                y: 0
+                            }), t.menu.removeClass(D), t.links.removeClass(L), t.overlay && t.overlay.children()
+                            .length && (R.length ? t.menu.insertAfter(R) : t.menu.prependTo(t.parent), t.overlay
+                                .attr("style", "").hide()), t.el.triggerHandler("w-close")
+                    }
+                    if (t.open) {
+                        t.open = !1, t.button.removeClass($);
+                        var i = t.config;
+                        if (("none" === i.animation || !E.support.transform || i.duration <= 0) && (e = !0),
+                            C.outro(0, t.el[0]), A.off("tap" + M, t.outside), e) return E(t.menu).stop(),
+                            void n();
+                        var r = "transform " + i.duration + "ms " + i.easing2,
+                            o = t.menu.outerHeight(!0),
+                            a = t.menu.outerWidth(!0),
+                            s = t.el.height();
+                        if (i.animOver) E(t.menu).add(r).start({
+                            x: a * i.animDirect
+                        }).then(n);
+                        else {
+                            var u = s + o;
+                            E(t.menu).add(r).start({
+                                y: -u
+                            }).then(n)
+                        }
+                    }
+                }
+                var _, k, T, O = {},
+                    E = t.tram,
+                    z = t(window),
+                    A = t(document),
+                    j = e.env(),
+                    q = '<div class="w-nav-overlay" data-wf-ignore />',
+                    M = ".w-nav",
+                    $ = "w--open",
+                    D = "w--nav-menu-open",
+                    L = "w--nav-link-open",
+                    C = n.triggers,
+                    R = t();
+                O.ready = O.design = O.preview = function() {
+                    T = j && e.env("design"), _ = t(document.body), (k = A.find(M)).length && (k.each(s),
+                        r(), o())
+                }, O.destroy = function() {
+                    R = t(), r(), k && k.length && k.each(u)
+                };
+                var B = "max-width";
+                return O
+            })
+        }), t(function(t) {
+            var e = o;
+            e.define("scroll", t.exports = function(t) {
+                function n(n, r) {
+                    if (l.test(n)) {
+                        var o = t("#" + n);
+                        if (o.length) {
+                            r && (r.preventDefault(), r.stopPropagation()), u.hash === n || !c || !c.pushState ||
+                                e.env.chrome && "file:" === u.protocol || (c.state && c.state.hash) !== n &&
+                                c.pushState({
+                                    hash: n
+                                }, "", "#" + n);
+                            var a = e.env("editor") ? ".w-editor-body" : "body",
+                                f = t("header, " + a + " > .header, " + a +
+                                    " > .w-nav:not([data-no-scroll])"),
+                                h = "fixed" === f.css("position") ? f.outerHeight() : 0;
+                            s.setTimeout(function() {
+                                i(o, h)
+                            }, r ? 0 : 300)
+                        }
+                    }
+                }
+
+                function i(e, n) {
+                    var i = t(s).scrollTop(),
+                        o = e.offset().top - n;
+                    if ("mid" === e.data("scroll")) {
+                        var a = t(s).height() - n,
+                            u = e.outerHeight();
+                        u < a && (o -= Math.round((a - u) / 2))
+                    }
+                    var c = 1;
+                    t("body").add(e).each(function() {
+                        var e = parseFloat(t(this).attr("data-scroll-time"), 10);
+                        !isNaN(e) && (0 === e || e > 0) && (c = e)
+                    }), Date.now || (Date.now = function() {
+                        return (new Date).getTime()
+                    });
+                    var l = Date.now(),
+                        f = s.requestAnimationFrame || s.mozRequestAnimationFrame || s.webkitRequestAnimationFrame ||
+                        function(t) {
+                            s.setTimeout(t, 15)
+                        },
+                        h = (472.143 * Math.log(Math.abs(i - o) + 125) - 2e3) * c,
+                        d = function() {
+                            var t = Date.now() - l;
+                            s.scroll(0, r(i, o, t, h)), t <= h && f(d)
+                        };
+                    d()
+                }
+
+                function r(t, e, n, i) {
+                    return n > i ? e : t + (e - t) * o(n / i)
+                }
+
+                function o(t) {
+                    return t < .5 ? 4 * t * t * t : (t - 1) * (2 * t - 2) * (2 * t - 2) + 1
+                }
+                var a = t(document),
+                    s = window,
+                    u = s.location,
+                    c = function() {
+                        try {
+                            return Boolean(s.frameElement)
+                        } catch (t) {
+                            return !0
+                        }
+                    }() ? null : s.history,
+                    l = /^[a-zA-Z0-9][\w:.-]*$/;
+                return {
+                    ready: function() {
+                        u.hash && n(u.hash.substring(1));
+                        var i = u.href.split("#")[0];
+                        a.on("click", "a", function(r) {
+                            if (!(e.env("design") || window.$.mobile && t(r.currentTarget).hasClass(
+                                    "ui-link")))
+                                if ("#" !== this.getAttribute("href")) {
+                                    var o = this.href.split("#"),
+                                        a = o[0] === i ? o[1] : null;
+                                    a && n(a, r)
+                                } else r.preventDefault()
                         })
                     }
                 }
-            }
-
-            function f(a) {
-                a = i(a, 0), this.active ? this.queue.push({options: a}) : (this.timer = new S({
-                    duration: a,
-                    context: this,
-                    complete: h
-                }), this.active = !0)
-            }
-
-            function g(a) {
-                return this.active ? (this.queue.push({
-                    options: a,
-                    args: arguments
-                }), void(this.timer.complete = h)) : j("No active transition timer. Use start() or wait() before then().")
-            }
-
-            function h() {
-                if (this.timer && this.timer.destroy(), this.active = !1, this.queue.length) {
-                    var a = this.queue.shift();
-                    e.call(this, a.options, !0, a.args)
-                }
-            }
-
-            function l(a) {
-                this.timer && this.timer.destroy(), this.queue = [], this.active = !1;
-                var b;
-                "string" == typeof a ? (b = {}, b[a] = 1) : b = "object" == typeof a && null != a ? a : this.props, u.call(this, b, v), t.call(this)
-            }
-
-            function m(a) {
-                l.call(this, a), u.call(this, a, w, x)
-            }
-
-            function n(a) {
-                "string" != typeof a && (a = "block"), this.el.style.display = a
-            }
-
-            function o() {
-                l.call(this), this.el.style.display = "none"
-            }
-
-            function p() {
-                this.el.offsetHeight
-            }
-
-            function r() {
-                l.call(this), a.removeData(this.el, q), this.$el = this.el = null
-            }
-
-            function t() {
-                var a, b, c = [];
-                this.upstream && c.push(this.upstream);
-                for (a in this.props)b = this.props[a], b.active && c.push(b.string);
-                c = c.join(","), this.style !== c && (this.style = c, this.el.style[G.transition.dom] = c)
-            }
-
-            function u(a, b, e) {
-                var f, g, h, i, j = b !== v, k = {};
-                for (f in a)h = a[f], f in Z ? (k.transform || (k.transform = {}), k.transform[f] = h) : (s.test(f) && (f = c(f)), f in Y ? k[f] = h : (i || (i = {}), i[f] = h));
-                for (f in k) {
-                    if (h = k[f], g = this.props[f], !g) {
-                        if (!j)continue;
-                        g = d.call(this, f)
-                    }
-                    b.call(this, g, h)
-                }
-                e && i && e.call(this, i)
-            }
-
-            function v(a) {
-                a.stop()
-            }
-
-            function w(a, b) {
-                a.set(b)
-            }
-
-            function x(a) {
-                this.$el.css(a)
-            }
-
-            function y(a, c) {
-                b[a] = function () {
-                    return this.children ? A.call(this, c, arguments) : (this.el && c.apply(this, arguments), this)
-                }
-            }
-
-            function A(a, b) {
-                var c, d = this.children.length;
-                for (c = 0; d > c; c++)a.apply(this.children[c], b);
-                return this
-            }
-
-            b.init = function (b) {
-                if (this.$el = a(b), this.el = this.$el[0], this.props = {}, this.queue = [], this.style = "", this.active = !1, U.keepInherited && !U.fallback) {
-                    var c = W(this.el, "transition");
-                    c && !z.test(c) && (this.upstream = c)
-                }
-                G.backface && U.hideBackface && V(this.el, G.backface.css, "hidden")
-            }, y("add", d), y("start", e), y("wait", f), y("then", g), y("next", h), y("stop", l), y("set", m), y("show", n), y("hide", o), y("redraw", p), y("destroy", r)
-        }), M = l(L, function (b) {
-            function c(b, c) {
-                var d = a.data(b, q) || a.data(b, q, new L.Bare);
-                return d.el || d.init(b), c ? d.start(c) : d
-            }
-
-            b.init = function (b, d) {
-                var e = a(b);
-                if (!e.length)return this;
-                if (1 === e.length)return c(e[0], d);
-                var f = [];
-                return e.each(function (a, b) {
-                    f.push(c(b, d))
-                }), this.children = f, this
-            }
-        }), N = l(function (a) {
-            function b() {
-                var a = this.get();
-                this.update("auto");
-                var b = this.get();
-                return this.update(a), b
-            }
-
-            function c(a, b, c) {
-                return void 0 !== b && (c = b), a in m ? a : c
-            }
-
-            function d(a) {
-                var b = /rgba?\((\d+),\s*(\d+),\s*(\d+)/.exec(a);
-                return (b ? e(b[1], b[2], b[3]) : a).replace(/#(\w)(\w)(\w)$/, "#$1$1$2$2$3$3")
-            }
-
-            var f = {duration: 500, ease: "ease", delay: 0};
-            a.init = function (a, b, d, e) {
-                this.$el = a, this.el = a[0];
-                var g = b[0];
-                d[2] && (g = d[2]), X[g] && (g = X[g]), this.name = g, this.type = d[1], this.duration = i(b[1], this.duration, f.duration), this.ease = c(b[2], this.ease, f.ease), this.delay = i(b[3], this.delay, f.delay), this.span = this.duration + this.delay, this.active = !1, this.nextStyle = null, this.auto = A.test(this.name), this.unit = e.unit || this.unit || U.defaultUnit, this.angle = e.angle || this.angle || U.defaultAngle, U.fallback || e.fallback ? this.animate = this.fallback : (this.animate = this.transition, this.string = this.name + B + this.duration + "ms" + ("ease" != this.ease ? B + m[this.ease][0] : "") + (this.delay ? B + this.delay + "ms" : ""))
-            }, a.set = function (a) {
-                a = this.convert(a, this.type), this.update(a), this.redraw()
-            }, a.transition = function (a) {
-                this.active = !0, a = this.convert(a, this.type), this.auto && ("auto" == this.el.style[this.name] && (this.update(this.get()), this.redraw()), "auto" == a && (a = b.call(this))), this.nextStyle = a
-            }, a.fallback = function (a) {
-                var c = this.el.style[this.name] || this.convert(this.get(), this.type);
-                a = this.convert(a, this.type), this.auto && ("auto" == c && (c = this.convert(this.get(), this.type)), "auto" == a && (a = b.call(this))), this.tween = new R({
-                    from: c,
-                    to: a,
-                    duration: this.duration,
-                    delay: this.delay,
-                    ease: this.ease,
-                    update: this.update,
-                    context: this
-                })
-            }, a.get = function () {
-                return W(this.el, this.name)
-            }, a.update = function (a) {
-                V(this.el, this.name, a)
-            }, a.stop = function () {
-                (this.active || this.nextStyle) && (this.active = !1, this.nextStyle = null, V(this.el, this.name, this.get()));
-                var a = this.tween;
-                a && a.context && a.destroy()
-            }, a.convert = function (a, b) {
-                if ("auto" == a && this.auto)return a;
-                var c, e = "number" == typeof a, f = "string" == typeof a;
-                switch (b) {
-                    case t:
-                        if (e)return a;
-                        if (f && "" === a.replace(r, ""))return +a;
-                        c = "number(unitless)";
-                        break;
-                    case u:
-                        if (f) {
-                            if ("" === a && this.original)return this.original;
-                            if (b.test(a))return "#" == a.charAt(0) && 7 == a.length ? a : d(a)
-                        }
-                        c = "hex or rgb string";
-                        break;
-                    case v:
-                        if (e)return a + this.unit;
-                        if (f && b.test(a))return a;
-                        c = "number(px) or string(unit)";
-                        break;
-                    case w:
-                        if (e)return a + this.unit;
-                        if (f && b.test(a))return a;
-                        c = "number(px) or string(unit or %)";
-                        break;
-                    case x:
-                        if (e)return a + this.angle;
-                        if (f && b.test(a))return a;
-                        c = "number(deg) or string(angle)";
-                        break;
-                    case y:
-                        if (e)return a;
-                        if (f && w.test(a))return a;
-                        c = "number(unitless) or string(unit or %)"
-                }
-                return g(c, a), a
-            }, a.redraw = function () {
-                this.el.offsetHeight
-            }
-        }), O = l(N, function (a, b) {
-            a.init = function () {
-                b.init.apply(this, arguments), this.original || (this.original = this.convert(this.get(), u))
-            }
-        }), P = l(N, function (a, b) {
-            a.init = function () {
-                b.init.apply(this, arguments), this.animate = this.fallback
-            }, a.get = function () {
-                return this.$el[this.name]()
-            }, a.update = function (a) {
-                this.$el[this.name](a)
-            }
-        }), Q = l(N, function (a, b) {
-            function c(a, b) {
-                var c, d, e, f, g;
-                for (c in a)f = Z[c], e = f[0], d = f[1] || c, g = this.convert(a[c], e), b.call(this, d, g, e)
-            }
-
-            a.init = function () {
-                b.init.apply(this, arguments), this.current || (this.current = {}, Z.perspective && U.perspective && (this.current.perspective = U.perspective, V(this.el, this.name, this.style(this.current)), this.redraw()))
-            }, a.set = function (a) {
-                c.call(this, a, function (a, b) {
-                    this.current[a] = b
-                }), V(this.el, this.name, this.style(this.current)), this.redraw()
-            }, a.transition = function (a) {
-                var b = this.values(a);
-                this.tween = new T({
-                    current: this.current,
-                    values: b,
-                    duration: this.duration,
-                    delay: this.delay,
-                    ease: this.ease
-                });
-                var c, d = {};
-                for (c in this.current)d[c] = c in b ? b[c] : this.current[c];
-                this.active = !0, this.nextStyle = this.style(d)
-            }, a.fallback = function (a) {
-                var b = this.values(a);
-                this.tween = new T({
-                    current: this.current,
-                    values: b,
-                    duration: this.duration,
-                    delay: this.delay,
-                    ease: this.ease,
-                    update: this.update,
-                    context: this
-                })
-            }, a.update = function () {
-                V(this.el, this.name, this.style(this.current))
-            }, a.style = function (a) {
-                var b, c = "";
-                for (b in a)c += b + "(" + a[b] + ") ";
-                return c
-            }, a.values = function (a) {
-                var b, d = {};
-                return c.call(this, a, function (a, c, e) {
-                    d[a] = c, void 0 === this.current[a] && (b = 0, ~a.indexOf("scale") && (b = 1), this.current[a] = this.convert(b, e))
-                }), d
-            }
-        }), R = l(function (b) {
-            function c(a) {
-                1 === n.push(a) && J(g)
-            }
-
-            function g() {
-                var a, b, c, d = n.length;
-                if (d)for (J(g), b = K(), a = d; a--;)c = n[a], c && c.render(b)
-            }
-
-            function i(b) {
-                var c, d = a.inArray(b, n);
-                d >= 0 && (c = n.slice(d + 1), n.length = d, c.length && (n = n.concat(c)))
-            }
-
-            function j(a) {
-                return Math.round(a * o) / o
-            }
-
-            function k(a, b, c) {
-                return e(a[0] + c * (b[0] - a[0]), a[1] + c * (b[1] - a[1]), a[2] + c * (b[2] - a[2]))
-            }
-
-            var l = {ease: m.ease[1], from: 0, to: 1};
-            b.init = function (a) {
-                this.duration = a.duration || 0, this.delay = a.delay || 0;
-                var b = a.ease || l.ease;
-                m[b] && (b = m[b][1]), "function" != typeof b && (b = l.ease), this.ease = b, this.update = a.update || f, this.complete = a.complete || f, this.context = a.context || this, this.name = a.name;
-                var c = a.from, d = a.to;
-                void 0 === c && (c = l.from), void 0 === d && (d = l.to), this.unit = a.unit || "", "number" == typeof c && "number" == typeof d ? (this.begin = c, this.change = d - c) : this.format(d, c), this.value = this.begin + this.unit, this.start = K(), a.autoplay !== !1 && this.play()
-            }, b.play = function () {
-                this.active || (this.start || (this.start = K()), this.active = !0, c(this))
-            }, b.stop = function () {
-                this.active && (this.active = !1, i(this))
-            }, b.render = function (a) {
-                var b, c = a - this.start;
-                if (this.delay) {
-                    if (c <= this.delay)return;
-                    c -= this.delay
-                }
-                if (c < this.duration) {
-                    var d = this.ease(c, 0, 1, this.duration);
-                    return b = this.startRGB ? k(this.startRGB, this.endRGB, d) : j(this.begin + d * this.change), this.value = b + this.unit, void this.update.call(this.context, this.value)
-                }
-                b = this.endHex || this.begin + this.change, this.value = b + this.unit, this.update.call(this.context, this.value), this.complete.call(this.context), this.destroy()
-            }, b.format = function (a, b) {
-                if (b += "", a += "", "#" == a.charAt(0))return this.startRGB = d(b), this.endRGB = d(a), this.endHex = a, this.begin = 0, void(this.change = 1);
-                if (!this.unit) {
-                    var c = b.replace(r, ""), e = a.replace(r, "");
-                    c !== e && h("tween", b, a), this.unit = c
-                }
-                b = parseFloat(b), a = parseFloat(a), this.begin = this.value = b, this.change = a - b
-            }, b.destroy = function () {
-                this.stop(), this.context = null, this.ease = this.update = this.complete = f
-            };
-            var n = [], o = 1e3
-        }), S = l(R, function (a) {
-            a.init = function (a) {
-                this.duration = a.duration || 0, this.complete = a.complete || f, this.context = a.context, this.play()
-            }, a.render = function (a) {
-                var b = a - this.start;
-                b < this.duration || (this.complete.call(this.context), this.destroy())
-            }
-        }), T = l(R, function (a, b) {
-            a.init = function (a) {
-                this.context = a.context, this.update = a.update, this.tweens = [], this.current = a.current;
-                var b, c;
-                for (b in a.values)c = a.values[b], this.current[b] !== c && this.tweens.push(new R({
-                    name: b,
-                    from: this.current[b],
-                    to: c,
-                    duration: a.duration,
-                    delay: a.delay,
-                    ease: a.ease,
-                    autoplay: !1
-                }));
-                this.play()
-            }, a.render = function (a) {
-                var b, c, d = this.tweens.length, e = !1;
-                for (b = d; b--;)c = this.tweens[b], c.context && (c.render(a), this.current[c.name] = c.value, e = !0);
-                return e ? void(this.update && this.update.call(this.context)) : this.destroy()
-            }, a.destroy = function () {
-                if (b.destroy.call(this), this.tweens) {
-                    var a, c = this.tweens.length;
-                    for (a = c; a--;)this.tweens[a].destroy();
-                    this.tweens = null, this.current = null
-                }
-            }
-        }), U = b.config = {
-            debug: !1,
-            defaultUnit: "px",
-            defaultAngle: "deg",
-            keepInherited: !1,
-            hideBackface: !1,
-            perspective: "",
-            fallback: !G.transition,
-            agentTests: []
-        };
-        b.fallback = function (a) {
-            if (!G.transition)return U.fallback = !0;
-            U.agentTests.push("(" + a + ")");
-            var b = new RegExp(U.agentTests.join("|"), "i");
-            U.fallback = b.test(navigator.userAgent)
-        }, b.fallback("6.0.[2-5] Safari"), b.tween = function (a) {
-            return new R(a)
-        }, b.delay = function (a, b, c) {
-            return new S({complete: b, duration: a, context: c})
-        }, a.fn.tram = function (a) {
-            return b.call(null, this, a)
-        };
-        var V = a.style, W = a.css, X = {transform: G.transform && G.transform.css}, Y = {
-            color: [O, u],
-            background: [O, u, "background-color"],
-            "outline-color": [O, u],
-            "border-color": [O, u],
-            "border-top-color": [O, u],
-            "border-right-color": [O, u],
-            "border-bottom-color": [O, u],
-            "border-left-color": [O, u],
-            "border-width": [N, v],
-            "border-top-width": [N, v],
-            "border-right-width": [N, v],
-            "border-bottom-width": [N, v],
-            "border-left-width": [N, v],
-            "border-spacing": [N, v],
-            "letter-spacing": [N, v],
-            margin: [N, v],
-            "margin-top": [N, v],
-            "margin-right": [N, v],
-            "margin-bottom": [N, v],
-            "margin-left": [N, v],
-            padding: [N, v],
-            "padding-top": [N, v],
-            "padding-right": [N, v],
-            "padding-bottom": [N, v],
-            "padding-left": [N, v],
-            "outline-width": [N, v],
-            opacity: [N, t],
-            top: [N, w],
-            right: [N, w],
-            bottom: [N, w],
-            left: [N, w],
-            "font-size": [N, w],
-            "text-indent": [N, w],
-            "word-spacing": [N, w],
-            width: [N, w],
-            "min-width": [N, w],
-            "max-width": [N, w],
-            height: [N, w],
-            "min-height": [N, w],
-            "max-height": [N, w],
-            "line-height": [N, y],
-            "scroll-top": [P, t, "scrollTop"],
-            "scroll-left": [P, t, "scrollLeft"]
-        }, Z = {};
-        G.transform && (Y.transform = [Q], Z = {
-            x: [w, "translateX"],
-            y: [w, "translateY"],
-            rotate: [x],
-            rotateX: [x],
-            rotateY: [x],
-            scale: [t],
-            scaleX: [t],
-            scaleY: [t],
-            skew: [x],
-            skewX: [x],
-            skewY: [x]
-        }), G.transform && G.backface && (Z.z = [w, "translateZ"], Z.rotateZ = [x], Z.scaleZ = [t], Z.perspective = [v]);
-        var $ = /ms/, _ = /s|\./;
-        return a.tram = b
-    }(window.jQuery);
-
-    var require$$0$1 = {};
-
-    var underscoreCustom = __commonjs(function (module) {
-        // Include tram for frame-throttling
-        var $ = window.$;
-        var tram = require$$0$1 && $.tram;
-
-        /*eslint-disable */
-
-        /*!
-         * Webflow._ (aka) Underscore.js 1.6.0 (custom build)
-         * _.each
-         * _.map
-         * _.find
-         * _.filter
-         * _.any
-         * _.contains
-         * _.delay
-         * _.defer
-         * _.throttle (webflow)
-         * _.debounce
-         * _.keys
-         * _.has
-         * _.now
-         *
-         * http://underscorejs.org
-         * (c) 2009-2013 Jeremy Ashkenas, DocumentCloud and Investigative Reporters & Editors
-         * Underscore may be freely distributed under the MIT license.
-         * @license MIT
-         */
-        module.exports = (function () {
-            var _ = {};
-
-            // Current version.
-            _.VERSION = '1.6.0-Webflow';
-
-            // Establish the object that gets returned to break out of a loop iteration.
-            var breaker = {};
-
-            // Save bytes in the minified (but not gzipped) version:
-            var ArrayProto = Array.prototype, ObjProto = Object.prototype, FuncProto = Function.prototype;
-
-            // Create quick reference variables for speed access to core prototypes.
-            var
-                push = ArrayProto.push,
-                slice = ArrayProto.slice,
-                concat = ArrayProto.concat,
-                toString = ObjProto.toString,
-                hasOwnProperty = ObjProto.hasOwnProperty;
-
-            // All **ECMAScript 5** native function implementations that we hope to use
-            // are declared here.
-            var
-                nativeForEach = ArrayProto.forEach,
-                nativeMap = ArrayProto.map,
-                nativeReduce = ArrayProto.reduce,
-                nativeReduceRight = ArrayProto.reduceRight,
-                nativeFilter = ArrayProto.filter,
-                nativeEvery = ArrayProto.every,
-                nativeSome = ArrayProto.some,
-                nativeIndexOf = ArrayProto.indexOf,
-                nativeLastIndexOf = ArrayProto.lastIndexOf,
-                nativeIsArray = Array.isArray,
-                nativeKeys = Object.keys,
-                nativeBind = FuncProto.bind;
-
-            // Collection Functions
-            // --------------------
-
-            // The cornerstone, an `each` implementation, aka `forEach`.
-            // Handles objects with the built-in `forEach`, arrays, and raw objects.
-            // Delegates to **ECMAScript 5**'s native `forEach` if available.
-            var each = _.each = _.forEach = function (obj, iterator, context) {
-                /* jshint shadow:true */
-                if (obj == null) return obj;
-                if (nativeForEach && obj.forEach === nativeForEach) {
-                    obj.forEach(iterator, context);
-                } else if (obj.length === +obj.length) {
-                    for (var i = 0, length = obj.length; i < length; i++) {
-                        if (iterator.call(context, obj[i], i, obj) === breaker) return;
-                    }
-                } else {
-                    var keys = _.keys(obj);
-                    for (var i = 0, length = keys.length; i < length; i++) {
-                        if (iterator.call(context, obj[keys[i]], keys[i], obj) === breaker) return;
-                    }
-                }
-                return obj;
-            };
-
-            // Return the results of applying the iterator to each element.
-            // Delegates to **ECMAScript 5**'s native `map` if available.
-            _.map = _.collect = function (obj, iterator, context) {
-                var results = [];
-                if (obj == null) return results;
-                if (nativeMap && obj.map === nativeMap) return obj.map(iterator, context);
-                each(obj, function (value, index, list) {
-                    results.push(iterator.call(context, value, index, list));
-                });
-                return results;
-            };
-
-            // Return the first value which passes a truth test. Aliased as `detect`.
-            _.find = _.detect = function (obj, predicate, context) {
-                var result;
-                any(obj, function (value, index, list) {
-                    if (predicate.call(context, value, index, list)) {
-                        result = value;
-                        return true;
-                    }
-                });
-                return result;
-            };
-
-            // Return all the elements that pass a truth test.
-            // Delegates to **ECMAScript 5**'s native `filter` if available.
-            // Aliased as `select`.
-            _.filter = _.select = function (obj, predicate, context) {
-                var results = [];
-                if (obj == null) return results;
-                if (nativeFilter && obj.filter === nativeFilter) return obj.filter(predicate, context);
-                each(obj, function (value, index, list) {
-                    if (predicate.call(context, value, index, list)) results.push(value);
-                });
-                return results;
-            };
-
-            // Determine if at least one element in the object matches a truth test.
-            // Delegates to **ECMAScript 5**'s native `some` if available.
-            // Aliased as `any`.
-            var any = _.some = _.any = function (obj, predicate, context) {
-                predicate || (predicate = _.identity);
-                var result = false;
-                if (obj == null) return result;
-                if (nativeSome && obj.some === nativeSome) return obj.some(predicate, context);
-                each(obj, function (value, index, list) {
-                    if (result || (result = predicate.call(context, value, index, list))) return breaker;
-                });
-                return !!result;
-            };
-
-            // Determine if the array or object contains a given value (using `===`).
-            // Aliased as `include`.
-            _.contains = _.include = function (obj, target) {
-                if (obj == null) return false;
-                if (nativeIndexOf && obj.indexOf === nativeIndexOf) return obj.indexOf(target) != -1;
-                return any(obj, function (value) {
-                    return value === target;
-                });
-            };
-
-            // Function (ahem) Functions
-            // --------------------
-
-            // Delays a function for the given number of milliseconds, and then calls
-            // it with the arguments supplied.
-            _.delay = function (func, wait) {
-                var args = slice.call(arguments, 2);
-                return setTimeout(function () {
-                    return func.apply(null, args);
-                }, wait);
-            };
-
-            // Defers a function, scheduling it to run after the current call stack has
-            // cleared.
-            _.defer = function (func) {
-                return _.delay.apply(_, [func, 1].concat(slice.call(arguments, 1)));
-            };
-
-            // Returns a function, that, when invoked, will only be triggered once every
-            // browser animation frame - using tram's requestAnimationFrame polyfill.
-            _.throttle = function (func) {
-                var wait, args, context;
-                return function () {
-                    if (wait) return;
-                    wait = true;
-                    args = arguments;
-                    context = this;
-                    tram.frame(function () {
-                        wait = false;
-                        func.apply(context, args);
-                    });
-                };
-            };
-
-            // Returns a function, that, as long as it continues to be invoked, will not
-            // be triggered. The function will be called after it stops being called for
-            // N milliseconds. If `immediate` is passed, trigger the function on the
-            // leading edge, instead of the trailing.
-            _.debounce = function (func, wait, immediate) {
-                var timeout, args, context, timestamp, result;
-
-                var later = function () {
-                    var last = _.now() - timestamp;
-                    if (last < wait) {
-                        timeout = setTimeout(later, wait - last);
-                    } else {
-                        timeout = null;
-                        if (!immediate) {
-                            result = func.apply(context, args);
-                            context = args = null;
-                        }
-                    }
-                };
-
-                return function () {
-                    context = this;
-                    args = arguments;
-                    timestamp = _.now();
-                    var callNow = immediate && !timeout;
-                    if (!timeout) {
-                        timeout = setTimeout(later, wait);
-                    }
-                    if (callNow) {
-                        result = func.apply(context, args);
-                        context = args = null;
-                    }
-
-                    return result;
-                };
-            };
-
-            // Object Functions
-            // ----------------
-
-            // Fill in a given object with default properties.
-            _.defaults = function (obj) {
-                if (!_.isObject(obj)) return obj;
-                for (var i = 1, length = arguments.length; i < length; i++) {
-                    var source = arguments[i];
-                    for (var prop in source) {
-                        if (obj[prop] === void 0) obj[prop] = source[prop];
-                    }
-                }
-                return obj;
-            };
-
-            // Retrieve the names of an object's properties.
-            // Delegates to **ECMAScript 5**'s native `Object.keys`
-            _.keys = function (obj) {
-                if (!_.isObject(obj)) return [];
-                if (nativeKeys) return nativeKeys(obj);
-                var keys = [];
-                for (var key in obj) if (_.has(obj, key)) keys.push(key);
-                return keys;
-            };
-
-            // Shortcut function for checking if an object has a given property directly
-            // on itself (in other words, not on a prototype).
-            _.has = function (obj, key) {
-                return hasOwnProperty.call(obj, key);
-            };
-
-            // Is a given variable an object?
-            _.isObject = function (obj) {
-                return obj === Object(obj);
-            };
-
-            // Utility Functions
-            // -----------------
-
-            // A (possibly faster) way to get the current timestamp as an integer.
-            _.now = Date.now || function () {
-                    return new Date().getTime();
-                };
-
-            // By default, Underscore uses ERB-style template delimiters, change the
-            // following template settings to use alternative delimiters.
-            _.templateSettings = {
-                evaluate: /<%([\s\S]+?)%>/g,
-                interpolate: /<%=([\s\S]+?)%>/g,
-                escape: /<%-([\s\S]+?)%>/g
-            };
-
-            // When customizing `templateSettings`, if you don't want to define an
-            // interpolation, evaluation or escaping regex, we need one that is
-            // guaranteed not to match.
-            var noMatch = /(.)^/;
-
-            // Certain characters need to be escaped so that they can be put into a
-            // string literal.
-            var escapes = {
-                "'": "'",
-                '\\': '\\',
-                '\r': 'r',
-                '\n': 'n',
-                '\u2028': 'u2028',
-                '\u2029': 'u2029'
-            };
-
-            var escaper = /\\|'|\r|\n|\u2028|\u2029/g;
-
-            var escapeChar = function (match) {
-                return '\\' + escapes[match];
-            };
-
-            // JavaScript micro-templating, similar to John Resig's implementation.
-            // Underscore templating handles arbitrary delimiters, preserves whitespace,
-            // and correctly escapes quotes within interpolated code.
-            // NB: `oldSettings` only exists for backwards compatibility.
-            _.template = function (text, settings, oldSettings) {
-                if (!settings && oldSettings) settings = oldSettings;
-                settings = _.defaults({}, settings, _.templateSettings);
-
-                // Combine delimiters into one regular expression via alternation.
-                var matcher = RegExp([
-                        (settings.escape || noMatch).source,
-                        (settings.interpolate || noMatch).source,
-                        (settings.evaluate || noMatch).source
-                    ].join('|') + '|$', 'g');
-
-                // Compile the template source, escaping string literals appropriately.
-                var index = 0;
-                var source = "__p+='";
-                text.replace(matcher, function (match, escape, interpolate, evaluate, offset) {
-                    source += text.slice(index, offset).replace(escaper, escapeChar);
-                    index = offset + match.length;
-
-                    if (escape) {
-                        source += "'+\n((__t=(" + escape + "))==null?'':_.escape(__t))+\n'";
-                    } else if (interpolate) {
-                        source += "'+\n((__t=(" + interpolate + "))==null?'':__t)+\n'";
-                    } else if (evaluate) {
-                        source += "';\n" + evaluate + "\n__p+='";
-                    }
-
-                    // Adobe VMs need the match returned to produce the correct offest.
-                    return match;
-                });
-                source += "';\n";
-
-                // If a variable is not specified, place data values in local scope.
-                if (!settings.variable) source = 'with(obj||{}){\n' + source + '}\n';
-
-                source = "var __t,__p='',__j=Array.prototype.join," +
-                    "print=function(){__p+=__j.call(arguments,'');};\n" +
-                    source + 'return __p;\n';
-
-                try {
-                    var render = new Function(settings.variable || 'obj', '_', source);
-                } catch (e) {
-                    e.source = source;
-                    throw e;
-                }
-
-                var template = function (data) {
-                    return render.call(this, data, _);
-                };
-
-                // Provide the compiled source as a convenience for precompilation.
-                var argument = settings.variable || 'obj';
-                template.source = 'function(' + argument + '){\n' + source + '}';
-
-                return template;
-            };
-
-            // Export underscore
-            return _;
-        }());
-
-        /* eslint-enable */
-    });
-
-    var require$$1 = (underscoreCustom && typeof underscoreCustom === 'object' && 'default' in underscoreCustom ? underscoreCustom['default'] : underscoreCustom);
-
-    var webflowLib = __commonjs(function (module) {
-        /**
-         * Webflow: Core site library
-         */
-
-        var Webflow = {};
-        var modules = {};
-        var primary = [];
-        var secondary = window.Webflow || [];
-        var $ = window.jQuery;
-        var $win = $(window);
-        var $doc = $(document);
-        var isFunction = $.isFunction;
-        var _ = Webflow._ = require$$1;
-        var tram = require$$0$1 && $.tram;
-        var domready = false;
-        var destroyed = false;
-        tram.config.hideBackface = false;
-        tram.config.keepInherited = true;
-
-        /**
-         * Webflow.define - Define a named module
-         * @param  {string} name
-         * @param  {function} factory
-         * @param  {object} options
-         * @return {object}
-         */
-        Webflow.define = function (name, factory, options) {
-            if (modules[name]) unbindModule(modules[name]);
-            var instance = modules[name] = factory($, _, options) || {};
-            bindModule(instance);
-            return instance;
-        };
-
-        /**
-         * Webflow.require - Require a named module
-         * @param  {string} name
-         * @return {object}
-         */
-        Webflow.require = function (name) {
-            return modules[name];
-        };
-
-        function bindModule(module) {
-            // If running in Webflow app, subscribe to design/preview events
-            if (Webflow.env()) {
-                isFunction(module.design) && $win.on('__wf_design', module.design);
-                isFunction(module.preview) && $win.on('__wf_preview', module.preview);
-            }
-            // Subscribe to front-end destroy event
-            isFunction(module.destroy) && $win.on('__wf_destroy', module.destroy);
-            // Look for ready method on module
-            if (module.ready && isFunction(module.ready)) {
-                addReady(module);
-            }
-        }
-
-        function addReady(module) {
-            // If domready has already happened, run ready method
-            if (domready) {
-                module.ready();
-                return;
-            }
-            // Otherwise add ready method to the primary queue (only once)
-            if (_.contains(primary, module.ready)) return;
-            primary.push(module.ready);
-        }
-
-        function unbindModule(module) {
-            // Unsubscribe module from window events
-            isFunction(module.design) && $win.off('__wf_design', module.design);
-            isFunction(module.preview) && $win.off('__wf_preview', module.preview);
-            isFunction(module.destroy) && $win.off('__wf_destroy', module.destroy);
-            // Remove ready method from primary queue
-            if (module.ready && isFunction(module.ready)) {
-                removeReady(module);
-            }
-        }
-
-        function removeReady(module) {
-            primary = _.filter(primary, function (readyFn) {
-                return readyFn !== module.ready;
-            });
-        }
-
-        /**
-         * Webflow.push - Add a ready handler into secondary queue
-         * @param {function} ready  Callback to invoke on domready
-         */
-        Webflow.push = function (ready) {
-            // If domready has already happened, invoke handler
-            if (domready) {
-                isFunction(ready) && ready();
-                return;
-            }
-            // Otherwise push into secondary queue
-            secondary.push(ready);
-        };
-
-        /**
-         * Webflow.env - Get the state of the Webflow app
-         * @param {string} mode [optional]
-         * @return {boolean}
-         */
-        Webflow.env = function (mode) {
-            var designFlag = window.__wf_design;
-            var inApp = typeof designFlag !== 'undefined';
-            if (!mode) return inApp;
-            if (mode === 'design') return inApp && designFlag;
-            if (mode === 'preview') return inApp && !designFlag;
-            if (mode === 'slug') return inApp && window.__wf_slug;
-            if (mode === 'editor') return window.WebflowEditor;
-            if (mode === 'test') return false || window.__wf_test;
-            if (mode === 'frame') return window !== window.top;
-        };
-
-        // Feature detects + browser sniffs  ಠ_ಠ
-        var userAgent = navigator.userAgent.toLowerCase();
-        var appVersion = navigator.appVersion.toLowerCase();
-        var touch = Webflow.env.touch = ('ontouchstart' in window) || window.DocumentTouch && document instanceof window.DocumentTouch;
-        var chrome = Webflow.env.chrome = /chrome/.test(userAgent) && /Google/.test(navigator.vendor) && parseInt(appVersion.match(/chrome\/(\d+)\./)[1], 10);
-        var ios = Webflow.env.ios = /(ipod|iphone|ipad)/.test(userAgent);
-        Webflow.env.safari = /safari/.test(userAgent) && !chrome && !ios;
-
-        // Maintain current touch target to prevent late clicks on touch devices
-        var touchTarget;
-        // Listen for both events to support touch/mouse hybrid devices
-        touch && $doc.on('touchstart mousedown', function (evt) {
-            touchTarget = evt.target;
-        });
-
-        /**
-         * Webflow.validClick - validate click target against current touch target
-         * @param  {HTMLElement} clickTarget  Element being clicked
-         * @return {Boolean}  True if click target is valid (always true on non-touch)
-         */
-        Webflow.validClick = touch ? function (clickTarget) {
-            return clickTarget === touchTarget || $.contains(clickTarget, touchTarget);
-        } : function () {
-            return true;
-        };
-
-        /**
-         * Webflow.resize, Webflow.scroll - throttled event proxies
-         */
-        var resizeEvents = 'resize.webflow orientationchange.webflow load.webflow';
-        var scrollEvents = 'scroll.webflow ' + resizeEvents;
-        Webflow.resize = eventProxy($win, resizeEvents);
-        Webflow.scroll = eventProxy($win, scrollEvents);
-        Webflow.redraw = eventProxy();
-
-        // Create a proxy instance for throttled events
-        function eventProxy(target, types) {
-
-            // Set up throttled method (using custom frame-based _.throttle)
-            var handlers = [];
-            var proxy = {};
-            proxy.up = _.throttle(function (evt) {
-                _.each(handlers, function (h) {
-                    h(evt);
-                });
-            });
-
-            // Bind events to target
-            if (target && types) target.on(types, proxy.up);
-
-            /**
-             * Add an event handler
-             * @param  {function} handler
-             */
-            proxy.on = function (handler) {
-                if (typeof handler !== 'function') return;
-                if (_.contains(handlers, handler)) return;
-                handlers.push(handler);
-            };
-
-            /**
-             * Remove an event handler
-             * @param  {function} handler
-             */
-            proxy.off = function (handler) {
-                // If no arguments supplied, clear all handlers
-                if (!arguments.length) {
-                    handlers = [];
-                    return;
-                }
-                // Otherwise, remove handler from the list
-                handlers = _.filter(handlers, function (h) {
-                    return h !== handler;
-                });
-            };
-
-            return proxy;
-        }
-
-        // Webflow.location - Wrap window.location in api
-        Webflow.location = function (url) {
-            window.location = url;
-        };
-
-        if (Webflow.env()) {
-            // Webflow.location - Re-route location change to trigger an event
-            Webflow.location = function (url) {
-                window.dispatchEvent(new CustomEvent('__wf_location', {detail: url}));
-            };
-        }
-
-        // Webflow.ready - Call primary and secondary handlers
-        Webflow.ready = function () {
-            domready = true;
-
-            // Restore modules after destroy
-            if (destroyed) {
-                restoreModules();
-
-                // Otherwise run primary ready methods
-            } else {
-                _.each(primary, callReady);
-            }
-
-            // Run secondary ready methods
-            _.each(secondary, callReady);
-
-            // Trigger resize
-            Webflow.resize.up();
-        };
-
-        function callReady(readyFn) {
-            isFunction(readyFn) && readyFn();
-        }
-
-        function restoreModules() {
-            destroyed = false;
-            _.each(modules, bindModule);
-        }
-
-        /**
-         * Webflow.load - Add a window load handler that will run even if load event has already happened
-         * @param  {function} handler
-         */
-        var deferLoad;
-        Webflow.load = function (handler) {
-            deferLoad.then(handler);
-        };
-
-        function bindLoad() {
-            // Reject any previous deferred (to support destroy)
-            if (deferLoad) {
-                deferLoad.reject();
-                $win.off('load', deferLoad.resolve);
-            }
-            // Create deferred and bind window load event
-            deferLoad = new $.Deferred();
-            $win.on('load', deferLoad.resolve);
-        }
-
-        // Webflow.destroy - Trigger a destroy event for all modules
-        Webflow.destroy = function (options) {
-            options = options || {};
-            destroyed = true;
-            $win.triggerHandler('__wf_destroy');
-
-            // Allow domready reset for tests
-            if (options.domready != null) {
-                domready = options.domready;
-            }
-
-            // Unbind modules
-            _.each(modules, unbindModule);
-
-            // Clear any proxy event handlers
-            Webflow.resize.off();
-            Webflow.scroll.off();
-            Webflow.redraw.off();
-
-            // Clear any queued ready methods
-            primary = [];
-            secondary = [];
-
-            // If load event has not yet fired, replace the deferred
-            if (deferLoad.state() === 'pending') bindLoad();
-        };
-
-        // Listen for domready
-        $(Webflow.ready);
-
-        // Listen for window.onload and resolve deferred
-        bindLoad();
-
-        // Export commonjs module
-        module.exports = window.Webflow = Webflow;
-    });
-
-    var require$$0 = (webflowLib && typeof webflowLib === 'object' && 'default' in webflowLib ? webflowLib['default'] : webflowLib);
-
-    var webflowBrand = __commonjs(function (module) {
-        /**
-         * Webflow: Brand pages on the subdomain
-         */
-
-        var Webflow = require$$0;
-
-        Webflow.define('brand', module.exports = function ($) {
-            var api = {};
-            var $html = $('html');
-            var $body = $('body');
-            var namespace = '.w-webflow-badge';
-            var location = window.location;
-            var isPhantom = /PhantomJS/i.test(navigator.userAgent);
-            var brandElement;
-
-            // -----------------------------------
-            // Module methods
-
-            api.ready = function () {
-                var shouldBrand = $html.attr('data-wf-status');
-                var publishedDomain = $html.attr('data-wf-domain') || '';
-                if (/\.webflow\.io$/i.test(publishedDomain) && location.hostname !== publishedDomain) {
-                    shouldBrand = true;
-                }
-                if (shouldBrand && !isPhantom) {
-                    brandElement = brandElement || createBadge();
-                    ensureBrand();
-                    setTimeout(ensureBrand, 500);
-                }
-            };
-
-            function createBadge() {
-                var $brand = $('<a class="w-webflow-badge"></a>')
-                    .attr('href', 'https://webflow.com?utm_campaign=brandjs');
-
-                var $logoArt = $('<img>')
-                    .attr('src', 'https://d1otoma47x30pg.cloudfront.net/img/webflow-badge-icon.60efbf6ec9.svg')
-                    .css({
-                        marginRight: '8px',
-                        width: '16px',
-                    });
-
-                var $logoText = $('<img>')
-                    .attr('src', 'https://d1otoma47x30pg.cloudfront.net/img/webflow-badge-text.6faa6a38cd.svg');
-
-                $brand.append($logoArt, $logoText);
-                return $brand[0];
-            }
-
-            function ensureBrand() {
-                var found = $body.children(namespace);
-                var match = found.length && found.get(0) === brandElement;
-                var inEditor = Webflow.env('editor');
-                if (match) {
-                    // Remove brand when Editor is active
-                    if (inEditor) {
-                        found.remove();
-                    }
-                    // Exit early, brand is in place
-                    return;
-                }
-                // Remove any invalid brand elements
-                if (found.length) {
-                    found.remove();
-                }
-                // Append the brand (unless Editor is active)
-                if (!inEditor) {
-                    $body.append(brandElement);
-                }
-            }
-
-            // Export module
-            return api;
-        });
-    });
-
-    var webflowIxEvents = __commonjs(function (module) {
-        'use strict';
-
-        /**
-         * Webflow: IX Event triggers for other modules
-         */
-
-        var $ = window.jQuery;
-        var api = {};
-        var eventQueue = [];
-        var namespace = '.w-ix';
-
-        var eventTriggers = {
-            reset: function (i, el) {
-                el.__wf_intro = null;
-            },
-            intro: function (i, el) {
-                if (el.__wf_intro) return;
-                el.__wf_intro = true;
-                $(el).triggerHandler(api.types.INTRO);
-            },
-            outro: function (i, el) {
-                if (!el.__wf_intro) return;
-                el.__wf_intro = null;
-                $(el).triggerHandler(api.types.OUTRO);
-            }
-        };
-
-        api.triggers = {};
-
-        api.types = {
-            INTRO: 'w-ix-intro' + namespace,
-            OUTRO: 'w-ix-outro' + namespace
-        };
-
-        // Trigger any events in queue + restore trigger methods
-        api.init = function () {
-            var count = eventQueue.length;
-            for (var i = 0; i < count; i++) {
-                var memo = eventQueue[i];
-                memo[0](0, memo[1]);
-            }
-            eventQueue = [];
-            $.extend(api.triggers, eventTriggers);
-        };
-
-        // Replace all triggers with async wrapper to queue events until init
-        api.async = function () {
-            for (var key in eventTriggers) {
-                var func = eventTriggers[key];
-                if (!eventTriggers.hasOwnProperty(key)) continue;
-
-                // Replace trigger method with async wrapper
-                api.triggers[key] = function (i, el) {
-                    eventQueue.push([func, el]);
-                };
-            }
-        };
-
-        // Default triggers to async queue
-        api.async();
-
-        module.exports = api;
-    });
-
-    var require$$0$2 = (webflowIxEvents && typeof webflowIxEvents === 'object' && 'default' in webflowIxEvents ? webflowIxEvents['default'] : webflowIxEvents);
-
-    var webflowDropdown = __commonjs(function (module) {
-        /**
-         * Webflow: Dropdown component
-         */
-
-        var Webflow = require$$0;
-        var IXEvents = require$$0$2;
-
-        Webflow.define('dropdown', module.exports = function ($, _) {
-            var api = {};
-            var $doc = $(document);
-            var $dropdowns;
-            var designer;
-            var inApp = Webflow.env();
-            var touch = Webflow.env.touch;
-            var namespace = '.w-dropdown';
-            var stateOpen = 'w--open';
-            var closeEvent = 'w-close' + namespace;
-            var ix = IXEvents.triggers;
-            var defaultZIndex = 900; // @dropdown-depth
-            var inPreview = false;
-
-            // -----------------------------------
-            // Module methods
-
-            api.ready = init;
-
-            api.design = function () {
-                // Close all when returning from preview
-                if (inPreview) {
-                    closeAll();
-                }
-                inPreview = false;
-                init();
-            };
-
-            api.preview = function () {
-                inPreview = true;
-                init();
-            };
-
-            // -----------------------------------
-            // Private methods
-
-            function init() {
-                designer = inApp && Webflow.env('design');
-
-                // Find all instances on the page
-                $dropdowns = $doc.find(namespace);
-                $dropdowns.each(build);
-            }
-
-            function build(i, el) {
-                var $el = $(el);
-
-                // Store state in data
-                var data = $.data(el, namespace);
-                if (!data) data = $.data(el, namespace, {open: false, el: $el, config: {}});
-                data.list = $el.children('.w-dropdown-list');
-                data.toggle = $el.children('.w-dropdown-toggle');
-                data.links = data.list.children('.w-dropdown-link');
-                data.outside = outside(data);
-                data.complete = complete(data);
-                data.leave = leave(data);
-
-                // Remove old events
-                $el.off(namespace);
-                data.toggle.off(namespace);
-
-                // Set config from data attributes
-                configure(data);
-
-                if (data.nav) data.nav.off(namespace);
-                data.nav = $el.closest('.w-nav');
-                data.nav.on(closeEvent, handler(data));
-
-                // Add events based on mode
-                if (designer) {
-                    $el.on('setting' + namespace, handler(data));
-                } else {
-                    data.toggle.on('tap' + namespace, toggle(data));
-                    if (data.config.hover) {
-                        data.toggle.on('mouseenter' + namespace, enter(data));
-                    }
-                    $el.on(closeEvent, handler(data));
-                    // Close in preview mode
-                    inApp && close(data);
-                }
-            }
-
-            function configure(data) {
-                // Determine if z-index should be managed
-                var zIndex = Number(data.el.css('z-index'));
-                data.manageZ = zIndex === defaultZIndex || zIndex === defaultZIndex + 1;
-
-                data.config = {
-                    hover: Boolean(data.el.attr('data-hover')) && !touch,
-                    delay: Number(data.el.attr('data-delay')) || 0
-                };
-            }
-
-            function handler(data) {
-                return function (evt, options) {
-                    options = options || {};
-
-                    if (evt.type === 'w-close') {
-                        return close(data);
-                    }
-
-                    if (evt.type === 'setting') {
-                        configure(data);
-                        options.open === true && open(data, true);
-                        options.open === false && close(data, true);
-                        return;
-                    }
-                };
-            }
-
-            function toggle(data) {
-                return _.debounce(function () {
-                    data.open ? close(data) : open(data);
-                });
-            }
-
-            function open(data) {
-                if (data.open) return;
-                closeOthers(data);
-                data.open = true;
-                data.list.addClass(stateOpen);
-                data.toggle.addClass(stateOpen);
-                ix.intro(0, data.el[0]);
-                Webflow.redraw.up();
-
-                // Increase z-index to keep above other managed dropdowns
-                data.manageZ && data.el.css('z-index', defaultZIndex + 1);
-
-                // Listen for tap outside events
-                if (!designer) $doc.on('tap' + namespace, data.outside);
-                if (data.hovering) data.el.on('mouseleave' + namespace, data.leave);
-
-                // Clear previous delay
-                window.clearTimeout(data.delayId);
-            }
-
-            function close(data, immediate) {
-                if (!data.open) return;
-
-                // Do not close hover-based menus if currently hovering
-                if (data.config.hover && data.hovering) return;
-
-                data.open = false;
-                var config = data.config;
-                ix.outro(0, data.el[0]);
-
-                // Stop listening for tap outside events
-                $doc.off('tap' + namespace, data.outside);
-                data.el.off('mouseleave' + namespace, data.leave);
-
-                // Clear previous delay
-                window.clearTimeout(data.delayId);
-
-                // Skip delay during immediate
-                if (!config.delay || immediate) return data.complete();
-
-                // Optionally wait for delay before close
-                data.delayId = window.setTimeout(data.complete, config.delay);
-            }
-
-            function closeAll() {
-                $doc.find(namespace).each(function (i, el) {
-                    $(el).triggerHandler(closeEvent);
-                });
-            }
-
-            function closeOthers(data) {
-                var self = data.el[0];
-                $dropdowns.each(function (i, other) {
-                    var $other = $(other);
-                    if ($other.is(self) || $other.has(self).length) return;
-                    $other.triggerHandler(closeEvent);
-                });
-            }
-
-            function outside(data) {
-                // Unbind previous tap handler if it exists
-                if (data.outside) $doc.off('tap' + namespace, data.outside);
-
-                // Close menu when tapped outside
-                return _.debounce(function (evt) {
-                    if (!data.open) return;
-                    var $target = $(evt.target);
-                    if ($target.closest('.w-dropdown-toggle').length) return;
-                    if (!data.el.is($target.closest(namespace))) {
-                        close(data);
-                    }
-                });
-            }
-
-            function complete(data) {
-                return function () {
-                    data.list.removeClass(stateOpen);
-                    data.toggle.removeClass(stateOpen);
-
-                    // Reset z-index for managed dropdowns
-                    data.manageZ && data.el.css('z-index', '');
-                };
-            }
-
-            function enter(data) {
-                return function () {
-                    data.hovering = true;
-                    open(data);
-                };
-            }
-
-            function leave(data) {
-                return function () {
-                    data.hovering = false;
-                    close(data);
-                };
-            }
-
-            // Export module
-            return api;
-        });
-    });
-
-    var xdomainMin = __commonjs(function (module) {
-        /*!
-         * jQuery-ajaxTransport-XDomainRequest - v1.0.3
-         * 2014-12-16 WEBFLOW - Removed UMD wrapper
-         * https://github.com/MoonScript/jQuery-ajaxTransport-XDomainRequest
-         * Copyright (c) 2014 Jason Moon (@JSONMOON)
-         * @license MIT (/blob/master/LICENSE.txt)
-         */
-        module.exports = function ($) {
-            if ($.support.cors || !$.ajaxTransport || !window.XDomainRequest) {
-                return
-            }
-            var httpRegEx = /^https?:\/\//i;
-            var getOrPostRegEx = /^get|post$/i;
-            var sameSchemeRegEx = new RegExp("^" + location.protocol, "i");
-            $.ajaxTransport("* text html xml json", function (options, userOptions, jqXHR) {
-                if (!options.crossDomain || !options.async || !getOrPostRegEx.test(options.type) || !httpRegEx.test(options.url) || !sameSchemeRegEx.test(options.url)) {
-                    return
-                }
-                var xdr = null;
-                return {
-                    send: function (headers, complete) {
-                        var postData = "";
-                        var userType = (userOptions.dataType || "").toLowerCase();
-                        xdr = new XDomainRequest;
-                        if (/^\d+$/.test(userOptions.timeout)) {
-                            xdr.timeout = userOptions.timeout
-                        }
-                        xdr.ontimeout = function () {
-                            complete(500, "timeout")
-                        };
-                        xdr.onload = function () {
-                            var allResponseHeaders = "Content-Length: " + xdr.responseText.length + "\r\nContent-Type: " + xdr.contentType;
-                            var status = {code: 200, message: "success"};
-                            var responses = {text: xdr.responseText};
-                            try {
-                                if (userType === "html" || /text\/html/i.test(xdr.contentType)) {
-                                    responses.html = xdr.responseText
-                                } else if (userType === "json" || userType !== "text" && /\/json/i.test(xdr.contentType)) {
-                                    try {
-                                        responses.json = $.parseJSON(xdr.responseText)
-                                    } catch (e) {
-                                        status.code = 500;
-                                        status.message = "parseerror"
-                                    }
-                                } else if (userType === "xml" || userType !== "text" && /\/xml/i.test(xdr.contentType)) {
-                                    var doc = new ActiveXObject("Microsoft.XMLDOM");
-                                    doc.async = false;
-                                    try {
-                                        doc.loadXML(xdr.responseText)
-                                    } catch (e) {
-                                        doc = undefined
-                                    }
-                                    if (!doc || !doc.documentElement || doc.getElementsByTagName("parsererror").length) {
-                                        status.code = 500;
-                                        status.message = "parseerror";
-                                        throw"Invalid XML: " + xdr.responseText
-                                    }
-                                    responses.xml = doc
-                                }
-                            } catch (parseMessage) {
-                                throw parseMessage
-                            } finally {
-                                complete(status.code, status.message, responses, allResponseHeaders)
-                            }
-                        };
-                        xdr.onprogress = function () {
-                        };
-                        xdr.onerror = function () {
-                            complete(500, "error", {text: xdr.responseText})
-                        };
-                        if (userOptions.data) {
-                            postData = $.type(userOptions.data) === "string" ? userOptions.data : $.param(userOptions.data)
-                        }
-                        xdr.open(options.type, options.url);
-                        xdr.send(postData)
-                    }, abort: function () {
-                        if (xdr) {
-                            xdr.abort()
-                        }
-                    }
-                }
             })
-        }(window.jQuery);
-    });
+        }), t(function(t) {
+            o.define("touch", t.exports = function(t) {
+                function e(t) {
+                    function e(t) {
+                        var e = t.touches;
+                        e && e.length > 1 || (l = !0, f = !1, e ? (h = !0, s = e[0].clientX, u = e[0].clientY) :
+                            (s = t.clientX, u = t.clientY), c = s)
+                    }
 
+                    function i(t) {
+                        if (l) {
+                            if (h && "mousemove" === t.type) return t.preventDefault(), void t.stopPropagation();
+                            var e = t.touches,
+                                i = e ? e[0].clientX : t.clientX,
+                                r = e ? e[0].clientY : t.clientY,
+                                p = i - c;
+                            c = i, Math.abs(p) > d && o && "" === String(o()) && (n("swipe", t, {
+                                direction: p > 0 ? "right" : "left"
+                            }), a()), (Math.abs(i - s) > 10 || Math.abs(r - u) > 10) && (f = !0)
+                        }
+                    }
 
-    var webflowLinks = __commonjs(function (module) {
-        /**
-         * Webflow: Auto-select links to current page or section
-         */
+                    function r(t) {
+                        if (l) {
+                            if (l = !1, h && "mouseup" === t.type) return t.preventDefault(), t.stopPropagation(),
+                                void(h = !1);
+                            f || n("tap", t)
+                        }
+                    }
 
-        var Webflow = require$$0;
-
-        Webflow.define('links', module.exports = function ($, _) {
-            var api = {};
-            var $win = $(window);
-            var designer;
-            var inApp = Webflow.env();
-            var location = window.location;
-            var tempLink = document.createElement('a');
-            var linkCurrent = 'w--current';
-            var validHash = /^#[a-zA-Z][\w:.-]*$/;
-            var indexPage = /index\.(html|php)$/;
-            var dirList = /\/$/;
-            var anchors;
-            var slug;
-
-            // -----------------------------------
-            // Module methods
-
-            api.ready = api.design = api.preview = init;
-
-            // -----------------------------------
-            // Private methods
-
-            function init() {
-                designer = inApp && Webflow.env('design');
-                slug = Webflow.env('slug') || location.pathname || '';
-
-                // Reset scroll listener, init anchors
-                Webflow.scroll.off(scroll);
-                anchors = [];
-
-                // Test all links for a selectable href
-                var links = document.links;
-                for (var i = 0; i < links.length; ++i) {
-                    select(links[i]);
+                    function a() {
+                        l = !1
+                    }
+                    var s, u, c, l = !1,
+                        f = !1,
+                        h = !1,
+                        d = Math.min(Math.round(.04 * window.innerWidth), 40);
+                    t.addEventListener("touchstart", e, !1), t.addEventListener("touchmove", i, !1), t.addEventListener(
+                            "touchend", r, !1), t.addEventListener("touchcancel", a, !1), t.addEventListener(
+                            "mousedown", e, !1), t.addEventListener("mousemove", i, !1), t.addEventListener(
+                            "mouseup", r, !1), t.addEventListener("mouseout", a, !1), this.destroy =
+                        function() {
+                            t.removeEventListener("touchstart", e, !1), t.removeEventListener("touchmove",
+                                i, !1), t.removeEventListener("touchend", r, !1), t.removeEventListener(
+                                "touchcancel", a, !1), t.removeEventListener("mousedown", e, !1), t.removeEventListener(
+                                "mousemove", i, !1), t.removeEventListener("mouseup", r, !1), t.removeEventListener(
+                                "mouseout", a, !1), t = null
+                        }
                 }
 
-                // Listen for scroll if any anchors exist
-                if (anchors.length) {
-                    Webflow.scroll.on(scroll);
-                    scroll();
-                }
-            }
-
-            function select(link) {
-                var href = (designer && link.getAttribute('href-disabled')) || link.getAttribute('href');
-                tempLink.href = href;
-
-                // Ignore any hrefs with a colon to safely avoid all uri schemes
-                if (href.indexOf(':') >= 0) return;
-
-                var $link = $(link);
-
-                // Check for valid hash links w/ sections and use scroll anchor
-                if (href.indexOf('#') === 0 && validHash.test(href)) {
-                    var $section = $(href);
-                    $section.length && anchors.push({link: $link, sec: $section, active: false});
-                    return;
-                }
-
-                // Ignore empty # links
-                if (href === '#') return;
-
-                // Determine whether the link should be selected
-                var match = (tempLink.href === location.href) || (href === slug) || (indexPage.test(href) && dirList.test(slug));
-                setClass($link, linkCurrent, match);
-            }
-
-            function scroll() {
-                var viewTop = $win.scrollTop();
-                var viewHeight = $win.height();
-
-                // Check each anchor for a section in view
-                _.each(anchors, function (anchor) {
-                    var $link = anchor.link;
-                    var $section = anchor.sec;
-                    var top = $section.offset().top;
-                    var height = $section.outerHeight();
-                    var offset = viewHeight * 0.5;
-                    var active = ($section.is(':visible') &&
-                    top + height - offset >= viewTop &&
-                    top + offset <= viewTop + viewHeight);
-                    if (anchor.active === active) return;
-                    anchor.active = active;
-                    setClass($link, linkCurrent, active);
-                    if (designer) $link[0].__wf_current = active;
-                });
-            }
-
-            function setClass($elem, className, add) {
-                var exists = $elem.hasClass(className);
-                if (add && exists) return;
-                if (!add && !exists) return;
-                add ? $elem.addClass(className) : $elem.removeClass(className);
-            }
-
-            // Export module
-            return api;
-        });
-    });
-
-    var webflowNavbar = __commonjs(function (module) {
-        /**
-         * Webflow: Navbar component
-         */
-
-        var Webflow = require$$0;
-        var IXEvents = require$$0$2;
-
-        Webflow.define('navbar', module.exports = function ($, _) {
-            var api = {};
-            var tram = $.tram;
-            var $win = $(window);
-            var $doc = $(document);
-            var $body;
-            var $navbars;
-            var designer;
-            var inApp = Webflow.env();
-            var overlay = '<div class="w-nav-overlay" data-wf-ignore />';
-            var namespace = '.w-nav';
-            var buttonOpen = 'w--open';
-            var menuOpen = 'w--nav-menu-open';
-            var linkOpen = 'w--nav-link-open';
-            var ix = IXEvents.triggers;
-            var menuSibling = $();
-
-            // -----------------------------------
-            // Module methods
-
-            api.ready = api.design = api.preview = init;
-
-            api.destroy = function () {
-                menuSibling = $();
-                removeListeners();
-                if ($navbars && $navbars.length) {
-                    $navbars.each(teardown);
-                }
-            };
-
-            // -----------------------------------
-            // Private methods
-
-            function init() {
-                designer = inApp && Webflow.env('design');
-                $body = $(document.body);
-
-                // Find all instances on the page
-                $navbars = $doc.find(namespace);
-                if (!$navbars.length) return;
-                $navbars.each(build);
-
-                // Wire events
-                removeListeners();
-                addListeners();
-            }
-
-            function removeListeners() {
-                Webflow.resize.off(resizeAll);
-            }
-
-            function addListeners() {
-                Webflow.resize.on(resizeAll);
-            }
-
-            function resizeAll() {
-                $navbars.each(resize);
-            }
-
-            function build(i, el) {
-                var $el = $(el);
-
-                // Store state in data
-                var data = $.data(el, namespace);
-                if (!data) data = $.data(el, namespace, {open: false, el: $el, config: {}});
-                data.menu = $el.find('.w-nav-menu');
-                data.links = data.menu.find('.w-nav-link');
-                data.dropdowns = data.menu.find('.w-dropdown');
-                data.button = $el.find('.w-nav-button');
-                data.container = $el.find('.w-container');
-                data.outside = outside(data);
-
-                // Remove old events
-                data.el.off(namespace);
-                data.button.off(namespace);
-                data.menu.off(namespace);
-
-                // Set config from data attributes
-                configure(data);
-
-                // Add events based on mode
-                if (designer) {
-                    removeOverlay(data);
-                    data.el.on('setting' + namespace, handler(data));
-                } else {
-                    addOverlay(data);
-                    data.button.on('tap' + namespace, toggle(data));
-                    data.menu.on('click' + namespace, 'a', navigate(data));
-                }
-
-                // Trigger initial resize
-                resize(i, el);
-            }
-
-            function teardown(i, el) {
-                var data = $.data(el, namespace);
-                if (data) {
-                    removeOverlay(data);
-                    $.removeData(el, namespace);
-                }
-            }
-
-            function removeOverlay(data) {
-                if (!data.overlay) return;
-                close(data, true);
-                data.overlay.remove();
-                data.overlay = null;
-            }
-
-            function addOverlay(data) {
-                if (data.overlay) return;
-                data.overlay = $(overlay).appendTo(data.el);
-                data.parent = data.menu.parent();
-                close(data, true);
-            }
-
-            function configure(data) {
-                var config = {};
-                var old = data.config || {};
-
-                // Set config options from data attributes
-                var animation = config.animation = data.el.attr('data-animation') || 'default';
-                config.animOver = /^over/.test(animation);
-                config.animDirect = /left$/.test(animation) ? -1 : 1;
-
-                // Re-open menu if the animation type changed
-                if (old.animation !== animation) {
-                    data.open && _.defer(reopen, data);
-                }
-
-                config.easing = data.el.attr('data-easing') || 'ease';
-                config.easing2 = data.el.attr('data-easing2') || 'ease';
-
-                var duration = data.el.attr('data-duration');
-                config.duration = duration != null ? Number(duration) : 400;
-
-                config.docHeight = data.el.attr('data-doc-height');
-
-                // Store config in data
-                data.config = config;
-            }
-
-            function handler(data) {
-                return function (evt, options) {
-                    options = options || {};
-                    var winWidth = $win.width();
-                    configure(data);
-                    options.open === true && open(data, true);
-                    options.open === false && close(data, true);
-                    // Reopen if media query changed after setting
-                    data.open && _.defer(function () {
-                        if (winWidth !== $win.width()) reopen(data);
+                function n(e, n, i) {
+                    var r = t.Event(e, {
+                        originalEvent: n
                     });
-                };
-            }
-
-            function reopen(data) {
-                if (!data.open) return;
-                close(data, true);
-                open(data, true);
-            }
-
-            function toggle(data) {
-                // Debounce toggle to wait for accurate open state
-                return _.debounce(function () {
-                    data.open ? close(data) : open(data);
-                });
-            }
-
-            function navigate(data) {
-                return function (evt) {
-                    var link = $(this);
-                    var href = link.attr('href');
-
-                    // Avoid late clicks on touch devices
-                    if (!Webflow.validClick(evt.currentTarget)) {
-                        evt.preventDefault();
-                        return;
-                    }
-
-                    // Close when navigating to an in-page anchor
-                    if (href && href.indexOf('#') === 0 && data.open) {
-                        close(data);
-                    }
-                };
-            }
-
-            function outside(data) {
-                // Unbind previous tap handler if it exists
-                if (data.outside) $doc.off('tap' + namespace, data.outside);
-
-                // Close menu when tapped outside, debounced to wait for state
-                return _.debounce(function (evt) {
-                    if (!data.open) return;
-                    var menu = $(evt.target).closest('.w-nav-menu');
-                    if (!data.menu.is(menu)) {
-                        close(data);
-                    }
-                });
-            }
-
-            function resize(i, el) {
-                var data = $.data(el, namespace);
-                // Check for collapsed state based on button display
-                var collapsed = data.collapsed = data.button.css('display') !== 'none';
-                // Close menu if button is no longer visible (and not in designer)
-                if (data.open && !collapsed && !designer) close(data, true);
-                // Set max-width of links + dropdowns to match container
-                if (data.container.length) {
-                    var updateEachMax = updateMax(data);
-                    data.links.each(updateEachMax);
-                    data.dropdowns.each(updateEachMax);
+                    t(n.target).trigger(r, i)
                 }
-                // If currently open, update height to match body
-                if (data.open) {
-                    setOverlayHeight(data);
-                }
-            }
-
-            var maxWidth = 'max-width';
-
-            function updateMax(data) {
-                // Set max-width of each element to match container
-                var containMax = data.container.css(maxWidth);
-                if (containMax === 'none') containMax = '';
-                return function (i, link) {
-                    link = $(link);
-                    link.css(maxWidth, '');
-                    // Don't set the max-width if an upstream value exists
-                    if (link.css(maxWidth) === 'none') link.css(maxWidth, containMax);
-                };
-            }
-
-            function open(data, immediate) {
-                if (data.open) return;
-                data.open = true;
-                data.menu.addClass(menuOpen);
-                data.links.addClass(linkOpen);
-                data.button.addClass(buttonOpen);
-                var config = data.config;
-                var animation = config.animation;
-                if (animation === 'none' || !tram.support.transform) immediate = true;
-                var bodyHeight = setOverlayHeight(data);
-                var menuHeight = data.menu.outerHeight(true);
-                var menuWidth = data.menu.outerWidth(true);
-                var navHeight = data.el.height();
-                var navbarEl = data.el[0];
-                resize(0, navbarEl);
-                ix.intro(0, navbarEl);
-                Webflow.redraw.up();
-
-                // Listen for tap outside events
-                if (!designer) $doc.on('tap' + namespace, data.outside);
-
-                // No transition for immediate
-                if (immediate) return;
-
-                var transConfig = 'transform ' + config.duration + 'ms ' + config.easing;
-
-                // Add menu to overlay
-                if (data.overlay) {
-                    menuSibling = data.menu.prev();
-                    data.overlay.show().append(data.menu);
-                }
-
-                // Over left/right
-                if (config.animOver) {
-                    tram(data.menu)
-                        .add(transConfig)
-                        .set({x: config.animDirect * menuWidth, height: bodyHeight}).start({x: 0});
-                    data.overlay && data.overlay.width(menuWidth);
-                    return;
-                }
-
-                // Drop Down
-                var offsetY = navHeight + menuHeight;
-                tram(data.menu)
-                    .add(transConfig)
-                    .set({y: -offsetY}).start({y: 0});
-            }
-
-            function setOverlayHeight(data) {
-                var config = data.config;
-                var bodyHeight = config.docHeight ? $doc.height() : $body.height();
-                if (config.animOver) {
-                    data.menu.height(bodyHeight);
-                } else if (data.el.css('position') !== 'fixed') {
-                    bodyHeight -= data.el.height();
-                }
-                data.overlay && data.overlay.height(bodyHeight);
-                return bodyHeight;
-            }
-
-            function close(data, immediate) {
-                if (!data.open) return;
-                data.open = false;
-                data.button.removeClass(buttonOpen);
-                var config = data.config;
-                if (config.animation === 'none' || !tram.support.transform || config.duration <= 0) immediate = true;
-                ix.outro(0, data.el[0]);
-
-                // Stop listening for tap outside events
-                $doc.off('tap' + namespace, data.outside);
-
-                if (immediate) {
-                    tram(data.menu).stop();
-                    complete();
-                    return;
-                }
-
-                var transConfig = 'transform ' + config.duration + 'ms ' + config.easing2;
-                var menuHeight = data.menu.outerHeight(true);
-                var menuWidth = data.menu.outerWidth(true);
-                var navHeight = data.el.height();
-
-                // Over left/right
-                if (config.animOver) {
-                    tram(data.menu)
-                        .add(transConfig)
-                        .start({x: menuWidth * config.animDirect}).then(complete);
-                    return;
-                }
-
-                // Drop Down
-                var offsetY = navHeight + menuHeight;
-                tram(data.menu)
-                    .add(transConfig)
-                    .start({y: -offsetY}).then(complete);
-
-                function complete() {
-                    data.menu.height('');
-                    tram(data.menu).set({x: 0, y: 0});
-                    data.menu.removeClass(menuOpen);
-                    data.links.removeClass(linkOpen);
-                    if (data.overlay && data.overlay.children().length) {
-                        // Move menu back to parent at the original location
-                        menuSibling.length ? data.menu.insertAfter(menuSibling) : data.menu.prependTo(data.parent);
-                        data.overlay.attr('style', '').hide();
-                    }
-
-                    // Trigger event so other components can hook in (dropdown)
-                    data.el.triggerHandler('w-close');
-                }
-            }
-
-            // Export module
-            return api;
-        });
-    });
-
-    var webflowScroll = __commonjs(function (module) {
-        /**
-         * Webflow: Smooth scroll
-         */
-
-        var Webflow = require$$0;
-
-        Webflow.define('scroll', module.exports = function ($) {
-            var $doc = $(document);
-            var win = window;
-            var loc = win.location;
-            var history = inIframe() ? null : win.history;
-            var validHash = /^[a-zA-Z0-9][\w:.-]*$/;
-
-            function inIframe() {
-                try {
-                    return Boolean(win.frameElement);
-                } catch (e) {
-                    return true;
-                }
-            }
-
-            function ready() {
-                // If hash is already present on page load, scroll to it right away
-                if (loc.hash) {
-                    findEl(loc.hash.substring(1));
-                }
-
-                // The current page url without the hash part.
-                var locHref = loc.href.split('#')[0];
-
-                // When clicking on a link, check if it links to another part of the page
-                $doc.on('click', 'a', function (e) {
-                    if (Webflow.env('design')) {
-                        return;
-                    }
-
-                    // Ignore links being used by jQuery mobile
-                    if (window.$.mobile && $(e.currentTarget).hasClass('ui-link')) return;
-
-                    // Ignore empty # links
-                    if (this.getAttribute('href') === '#') {
-                        e.preventDefault();
-                        return;
-                    }
-
-                    // The href property always contains the full url so we can compare
-                    // with the document’s location to only target links on this page.
-                    var parts = this.href.split('#');
-                    var hash = parts[0] === locHref ? parts[1] : null;
-                    if (hash) {
-                        findEl(hash, e);
-                    }
-                });
-            }
-
-            function findEl(hash, e) {
-                if (!validHash.test(hash)) return;
-
-                var el = $('#' + hash);
-                if (!el.length) {
-                    return;
-                }
-
-                if (e) {
-                    e.preventDefault();
-                    e.stopPropagation();
-                }
-
-                // Push new history state
-                if (
-                    loc.hash !== hash &&
-                    history && history.pushState &&
-                    // Navigation breaks Chrome when the protocol is `file:`.
-                    !(Webflow.env.chrome && loc.protocol === 'file:')
-                ) {
-                    var oldHash = history.state && history.state.hash;
-                    if (oldHash !== hash) {
-                        history.pushState({hash: hash}, '', '#' + hash);
-                    }
-                }
-
-                // If a fixed header exists, offset for the height
-                var rootTag = Webflow.env('editor') ? '.w-editor-body' : 'body';
-                var header = $('header, ' + rootTag + ' > .header, ' + rootTag + ' > .w-nav:not([data-no-scroll])');
-                var offset = header.css('position') === 'fixed' ? header.outerHeight() : 0;
-
-                win.setTimeout(function () {
-                    scroll(el, offset);
-                }, e ? 0 : 300);
-            }
-
-            function scroll(el, offset) {
-                var start = $(win).scrollTop();
-                var end = el.offset().top - offset;
-
-                // If specified, scroll so that the element ends up in the middle of the viewport
-                if (el.data('scroll') === 'mid') {
-                    var available = $(win).height() - offset;
-                    var elHeight = el.outerHeight();
-                    if (elHeight < available) {
-                        end -= Math.round((available - elHeight) / 2);
-                    }
-                }
-
-                var mult = 1;
-
-                // Check for custom time multiplier on the body and the element
-                $('body').add(el).each(function () {
-                    var time = parseFloat($(this).attr('data-scroll-time'), 10);
-                    if (!isNaN(time) && (time === 0 || time > 0)) {
-                        mult = time;
-                    }
-                });
-
-                // Shim for IE8 and below
-                if (!Date.now) {
-                    Date.now = function () {
-                        return new Date().getTime();
-                    };
-                }
-
-                var clock = Date.now();
-                var animate = win.requestAnimationFrame || win.mozRequestAnimationFrame || win.webkitRequestAnimationFrame || function (fn) {
-                        win.setTimeout(fn, 15);
-                    };
-                var duration = (472.143 * Math.log(Math.abs(start - end) + 125) - 2000) * mult;
-
-                var step = function () {
-                    var elapsed = Date.now() - clock;
-                    win.scroll(0, getY(start, end, elapsed, duration));
-
-                    if (elapsed <= duration) {
-                        animate(step);
-                    }
-                };
-
-                step();
-            }
-
-            function getY(start, end, elapsed, duration) {
-                if (elapsed > duration) {
-                    return end;
-                }
-
-                return start + (end - start) * ease(elapsed / duration);
-            }
-
-            function ease(t) {
-                return t < 0.5 ? 4 * t * t * t : (t - 1) * (2 * t - 2) * (2 * t - 2) + 1;
-            }
-
-            // Export module
-            return {ready: ready};
-        });
-    });
-
-    var webflowTouch = __commonjs(function (module) {
-        /**
-         * Webflow: Touch events
-         */
-
-        var Webflow = require$$0;
-
-        Webflow.define('touch', module.exports = function ($) {
-            var api = {};
-            var fallback = !document.addEventListener;
-            var getSelection = window.getSelection;
-
-            // Fallback to click events in old IE
-            if (fallback) {
-                $.event.special.tap = {bindType: 'click', delegateType: 'click'};
-            }
-
-            api.init = function (el) {
-                if (fallback) return null;
-                el = typeof el === 'string' ? $(el).get(0) : el;
-                return el ? new Touch(el) : null;
-            };
-
-            function Touch(el) {
-                var active = false;
-                var dirty = false;
-                var useTouch = false;
-                var thresholdX = Math.min(Math.round(window.innerWidth * 0.04), 40);
-                var startX;
-                var startY;
-                var lastX;
-
-                el.addEventListener('touchstart', start, false);
-                el.addEventListener('touchmove', move, false);
-                el.addEventListener('touchend', end, false);
-                el.addEventListener('touchcancel', cancel, false);
-                el.addEventListener('mousedown', start, false);
-                el.addEventListener('mousemove', move, false);
-                el.addEventListener('mouseup', end, false);
-                el.addEventListener('mouseout', cancel, false);
-
-                function start(evt) {
-                    // We don’t handle multi-touch events yet.
-                    var touches = evt.touches;
-                    if (touches && touches.length > 1) {
-                        return;
-                    }
-
-                    active = true;
-                    dirty = false;
-
-                    if (touches) {
-                        useTouch = true;
-                        startX = touches[0].clientX;
-                        startY = touches[0].clientY;
-                    } else {
-                        startX = evt.clientX;
-                        startY = evt.clientY;
-                    }
-
-                    lastX = startX;
-                }
-
-                function move(evt) {
-                    if (!active) return;
-
-                    if (useTouch && evt.type === 'mousemove') {
-                        evt.preventDefault();
-                        evt.stopPropagation();
-                        return;
-                    }
-
-                    var touches = evt.touches;
-                    var x = touches ? touches[0].clientX : evt.clientX;
-                    var y = touches ? touches[0].clientY : evt.clientY;
-
-                    var velocityX = x - lastX;
-                    lastX = x;
-
-                    // Allow swipes while pointer is down, but prevent them during text selection
-                    if (Math.abs(velocityX) > thresholdX && getSelection && String(getSelection()) === '') {
-                        triggerEvent('swipe', evt, {direction: velocityX > 0 ? 'right' : 'left'});
-                        cancel();
-                    }
-
-                    // If pointer moves more than 10px flag to cancel tap
-                    if (Math.abs(x - startX) > 10 || Math.abs(y - startY) > 10) {
-                        dirty = true;
-                    }
-                }
-
-                function end(evt) {
-                    if (!active) return;
-                    active = false;
-
-                    if (useTouch && evt.type === 'mouseup') {
-                        evt.preventDefault();
-                        evt.stopPropagation();
-                        useTouch = false;
-                        return;
-                    }
-
-                    if (!dirty) triggerEvent('tap', evt);
-                }
-
-                function cancel() {
-                    active = false;
-                }
-
-                function destroy() {
-                    el.removeEventListener('touchstart', start, false);
-                    el.removeEventListener('touchmove', move, false);
-                    el.removeEventListener('touchend', end, false);
-                    el.removeEventListener('touchcancel', cancel, false);
-                    el.removeEventListener('mousedown', start, false);
-                    el.removeEventListener('mousemove', move, false);
-                    el.removeEventListener('mouseup', end, false);
-                    el.removeEventListener('mouseout', cancel, false);
-                    el = null;
-                }
-
-                // Public instance methods
-                this.destroy = destroy;
-            }
-
-            // Wrap native event to supoprt preventdefault + stopPropagation
-            function triggerEvent(type, evt, data) {
-                var newEvent = $.Event(type, {originalEvent: evt});
-                $(evt.target).trigger(newEvent, data);
-            }
-
-            // Listen for touch events on all nodes by default.
-            api.instance = api.init(document);
-
-            // Export module
-            return api;
-        });
-    });
-
-}());
-
+                var i = {},
+                    r = !document.addEventListener,
+                    o = window.getSelection;
+                return r && (t.event.special.tap = {
+                    bindType: "click",
+                    delegateType: "click"
+                }), i.init = function(n) {
+                    return r ? null : (n = "string" == typeof n ? t(n).get(0) : n) ? new e(n) : null
+                }, i.instance = i.init(document), i
+            })
+        })
+}();
+/**
+ * ----------------------------------------------------------------------
+ * Webflow: Interactions: Init
+ */
+Webflow.require('ix').init([{
+    "slug": "show-ingredients",
+    "name": "Show ingredients",
+    "value": {
+        "style": {},
+        "triggers": [{
+            "type": "click",
+            "selector": ".ingredients-text",
+            "siblings": true,
+            "stepsA": [{
+                "wait": "100ms",
+                "opacity": 0.6,
+                "transition": "opacity 200 ease-in-out-quart 0, transform 200 ease 0"
+            }, {
+                "wait": "150ms",
+                "display": "block",
+                "opacity": 1,
+                "width": "100%",
+                "transition": "opacity 200 ease-in-out-quart 0, width 200 ease-in-out-quart 0, height 200 ease-in-out-quart 0, transform 200 ease 0"
+            }],
+            "stepsB": [{
+                "wait": "150ms",
+                "display": "none",
+                "transition": "transform 200 ease 0"
+            }]
+        }]
+    }
+}, {
+    "slug": "new-interaction",
+    "name": "New Interaction",
+    "value": {
+        "style": {},
+        "triggers": [{
+            "type": "click",
+            "selector": ".card-ingredients-container",
+            "siblings": true,
+            "stepsA": [{
+                "width": "100%",
+                "transition": "width 200 ease 0"
+            }],
+            "stepsB": [{
+                "width": "49%",
+                "transition": "width 200 ease 0"
+            }]
+        }]
+    }
+}]);
