+++ templates/head.html
@@ -2,17 +2,26 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>Accidentally vegan</title>
+    <title>Veganer kataloget</title>
+    <meta content="Veganer kataloget" property="og:title">
     <meta content="width=device-width, initial-scale=1" name="viewport">
-    <link rel=stylesheet type=text/css href="{{ url_for('static', filename='css/normalize.css') }}">
-    <link rel=stylesheet type=text/css href="{{ url_for('static', filename='css/webflow.css') }}">
-    <link rel=stylesheet type=text/css href="{{ url_for('static', filename='css/accidentally-vegan.webflow.css') }}">
-    <!-- [if lt IE 9]>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.min.js" type="text/javascript"></script>
+    <meta content="Webflow" name="generator">
+    <link href="css/normalize.css" rel="stylesheet" type="text/css">
+    <link href="css/webflow.css" rel="stylesheet" type="text/css">
+    <link href="css/item-catalog.webflow.css" rel="stylesheet" type="text/css">
+    <script src="https://ajax.googleapis.com/ajax/libs/webfont/1.4.7/webfont.js" type="text/javascript"></script>
+    <script type="text/javascript">WebFont.load({
+        google: {
+            families: ["Open Sans:300,300italic,400,400italic,600,600italic,700,700italic,800,800italic"]
+        }
+    });</script>
+    <!-- [if lt IE 9]><script src="https://cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.min.js" type="text/javascript"></script>
     <![endif] -->
-
-
+    <script type="text/javascript">!function (o, c) {
+        var n = c.documentElement, t = " w-mod-";
+        n.className += t + "js", ("ontouchstart" in o || o.DocumentTouch && c instanceof DocumentTouch) && (n.className += t + "touch")
+    }(window, document);
+    </script>
     <link href="https://daks2k3a4ib2z.cloudfront.net/img/favicon.ico" rel="shortcut icon" type="image/x-icon">
     <link href="https://daks2k3a4ib2z.cloudfront.net/img/webclip.png" rel="apple-touch-icon">
-
 </head>
