+++ templates/new-design/main.html
@@ -24,6 +24,7 @@
                         <a class="dropdown-text-block" href="/category">Kategorier</a>
                         <div class="dropdown-icon w-icon-dropdown-toggle"></div>
                     </div>
+
                     {% for category in menu_categories %}
 
                         <nav class="dropdown-list w-dropdown-list">
@@ -33,12 +34,14 @@
                         </nav>
 
                     {% endfor %}
+
                 </div>
                 <div class="dropdown w-dropdown" data-delay="0" data-hover="1">
                     <div class="nav-link-dropdown w-clearfix w-dropdown-toggle">
                         <a class="dropdown-text-block" href="{{ url_for('all_shops') }}">Butikker</a>
                         <div class="dropdown-icon w-icon-dropdown-toggle"></div>
                     </div>
+
                     {% for shop in menu_shops %}
 
                         <nav class="dropdown-list w-dropdown-list">
@@ -48,12 +51,14 @@
                         </nav>
 
                     {% endfor %}
+
                 </div>
                 <div class="dropdown w-dropdown" data-delay="0" data-hover="1">
                     <div class="nav-link-dropdown w-clearfix w-dropdown-toggle">
                         <a class="dropdown-text-block" href="{{ url_for('all_manufacturers') }}">Producenter</a>
                         <div class="dropdown-icon w-icon-dropdown-toggle"></div>
                     </div>
+
                     {% for manufacturer in menu_manufacturers %}
 
                         <nav class="dropdown-list w-dropdown-list">
@@ -63,9 +68,12 @@
                         </nav>
 
                     {% endfor %}
-                </div>
 
-                <a class="nav-help-btn w-nav-link" href="#">Hjælp</a></div>
+                </div>
+                <a class="nav-help-btn w-nav-link" href="{{ url_for('show_login') }}">Log ind</a>
+                <a class="nav-help-btn w-nav-link" href="{{ url_for('disconnect') }}">Log ud</a>
+                <a class="nav-help-btn w-nav-link" href="#">Hjælp</a>
+            </div>
         </nav>
         <div class="menu-button w-nav-button">
             <div class="icon w-icon-nav-menu"></div>
@@ -77,10 +85,15 @@
 
     {% with messages = get_flashed_messages() %}
         {% if messages %}
+
             <div class=flashes>
+
                 {% for message in messages %}
-                    {{ message }}
+
+                    <p>{{ message }}</p><br>
+
                 {% endfor %}
+
             </div>
         {% endif %}
     {% endwith %}
@@ -93,7 +106,7 @@
     </footer>
 </div>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.0/jquery.min.js" type="text/javascript"></script>
-<script src="js/webflow.js" type="text/javascript"></script>
+<script src="/static/js/webflow.js" type="text/javascript"></script>
 <!-- [if lte IE 9]><script src="https://cdnjs.cloudflare.com/ajax/libs/placeholders/3.0.2/placeholders.min.js"></script>
 <![endif] -->
 </body>
