+++ templates/new-design/new-entry.html
@@ -14,7 +14,7 @@
   <script src="https://ajax.googleapis.com/ajax/libs/webfont/1.4.7/webfont.js" type="text/javascript"></script>
   <script type="text/javascript">WebFont.load({
   google: {
-    families: ["Open Sans:300,300italic,400,400italic,600,600italic,700,700italic,800,800italic","PT Sans Caption:regular,700","PT Sans Narrow:regular,700"]
+    families: ["Open Sans:300,300italic,400,400italic,600,600italic,700,700italic,800,800italic"]
   }
 });</script>
   <!-- [if lt IE 9]><script src="https://cdnjs.cloudflare.com/ajax/libs/html5shiv/3.7.3/html5shiv.min.js" type="text/javascript"></script><![endif] -->
